% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALM.R
\name{ALM}
\alias{ALM}
\title{Augmented Lagrangian Minimizer}
\usage{
ALM(
  M,
  curv,
  nsteps = 10,
  plt = TRUE,
  method = "SLSQP",
  maxiter_solver = 100,
  ncores = 4,
  LAfreq = 15
)
}
\arguments{
\item{M}{The input membrane with initial data and reference}

\item{curv}{target curvature}

\item{nsteps}{number of MMC steps to be run}

\item{plt}{(boolean) for plotting}

\item{method}{"SLSQP" for internal nloptr minimizer}

\item{kTfreq}{frequency for cooling factor applied on kT}
}
\value{
membrane object with updates from MMC with data:

LA: list of recorded coefficients A

A: last coefficients

M.mu: last mu value

M.lam: last lambda vector

history: history of App-calls that created the result
}
\description{
run ALM on an existing membrane object to adjust shape to target curvature, minimizing the residual norm RN.
}
\examples{
M <- MakeStandardRBC(L=5)
plot(M)
M <- ALM(M,curv=86,20,plt=TRUE)
plot(M)
M
}
