% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Brechbuehler.Init.uv.2}
\alias{Brechbuehler.Init.uv.2}
\title{Brechbuehler.Init.uv.2}
\usage{
Brechbuehler.Init.uv.2(
  X1,
  Fit_order = 12,
  InitFit = FALSE,
  poles.axis = 2,
  mat.mode = c("cotan.lapl", "cotan.chi", "euclid"),
  file.out = "Brechbuehler-init-uv-2.obj",
  rxy = 0.01
)
}
\arguments{
\item{X1}{input 3d mesh object}

\item{Fit_order}{(=12) spectral order to fit, if InitFit=TRUE}

\item{InitFit}{(=FALSE)}

\item{poles.axis}{(=2) take North/Southpole on from this principal component (1...3); chose 4 for interactive picking by mouse}

\item{mat.mode}{: one from c("cotan.lapl", "cotan.chi", "euclid")}

\item{file.out}{(="Brechbuehler-init-uv-2.obj")}

\item{rxy}{(=0.01) tolerance parameter}
}
\value{
List containing uv (angles) and 3d mesh object O
}
\description{
computes spherical coordinates (u,v) for vertices of a 3D-object X
The original algorithm from Brechbühler (1995) is implemented,
but spherical areas are not iterated for refinement.
The resulting (u,v) may not be optimal for fitting.
A better result comes from algorithm GEMINI_disk_conformal_map()
}
