% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{CenterX}
\alias{CenterX}
\title{CenterX}
\usage{
CenterX(X)
}
\arguments{
\item{X}{input 3d coords}
}
\value{
centered coords
}
\description{
gives back centered coordinates X
ATTENTION: not center of mass, if X contains doubles!
}
