% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{Check_C0}
\alias{Check_C0}
\title{CheckC0}
\usage{
Check_C0(M)
}
\arguments{
\item{M}{membrane object to check}
}
\value{
-
}
\description{
check whether stored C0 equals global M.C0
}
\examples{
data("M4")
CheckC0(M4)
}
