% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNM.R
\name{CNM}
\alias{CNM}
\title{Constraint Newton Minimizer}
\usage{
CNM(
  M,
  nsteps = 5,
  del = 0.3,
  del_control = FALSE,
  diag.reg = 0,
  pinv = TRUE,
  LAfreq = 1,
  Mtol_Newton = 1e-04,
  ncores = 4,
  keepEig2A = FALSE,
  plt = TRUE,
  cluster = FALSE,
  use_serial_cpp = FALSE
)
}
\arguments{
\item{M}{The input membrane with initial data and reference}

\item{nsteps}{(=5) number of Newton steps to be performed}

\item{del}{(=0.3) update factor delte (<1)}

\item{LAfreq}{(=1) storage frequency into list of coefficients LA}

\item{Mtol_Newton}{(=1e-4) stopping criterion for KKT norm and constraints norm}

\item{ncores}{(=4) for parallel Hessian on ncores CPU-cores}

\item{plt}{(=TRUE) for plotting two figures with alpha/beta stress-shear parameters}

\item{cluster}{(=FALSE) for parallel Hessian using a cluster with ncores processes}

\item{keepEig}{(=FALSE) to keep eigensystem as attributes in A (and thus LA[[]] -> high memory demands)}
}
\value{
membrane object with updated data from CNM:

LA: list of recorded coefficients A

A: last coefficients

history: history of App-calls that created the result

E_CNM: vector of nsteps energy values

CNMiter: number of iterations, incl. previous calls

proc_time: aggregated processing time of all Apps called before with this object

Eig.H: eigensystem of last energy+constraints Hessian

Grad: final energy gradient

Jacobian: constraint Jacobian

Hessian: energy Hessian matrix

ConsHessians: Lagrangian Hessians from active constraints
}
\description{
run CNM on an existing membrane object to minimize energy under exact constraints on area and volume
You may include curvature constraint as well.
Temperature is controlled by parameter kT.
}
\examples{
data(L9__Stomatocyce_6); D5->M
M.C0=-6
plot(M)
M <- CNM(M,nsteps=10,cluster=TRUE)
plot(M)
M
}
