% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{color2d}
\alias{color2d}
\title{makes a 2d color array of size 2x2
with RGB colors on the corners of a rectangle
Not yet used, but could be for 2d colors encoding alpha and beta in a single color.}
\usage{
color2d(
  col1 = c(0, 0, 0),
  col2 = c(1, 0, 0),
  col3 = c(0, 0, 1),
  col4 = c(0, 1, 0),
  n = 25
)
}
\arguments{
\item{col1, col2, col3, col4}{rgb-vectors of corners colors, `col2=c(1,0,0)` for red.}

\item{n}{(=25) number of colors per side}
}
\value{
array of RGB colors from bilinear interpolation
}
\description{
makes a 2d color array of size 2x2
with RGB colors on the corners of a rectangle
Not yet used, but could be for 2d colors encoding alpha and beta in a single color.
}
