% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Cons_filter_delta}
\alias{Cons_filter_delta}
\title{Cons_filter_delta}
\usage{
Cons_filter_delta(delta, grd, bas, H2, nm = TRUE)
}
\arguments{
\item{H2}{current E_SCM}
}
\value{
projected delta as nx3 matrix
}
\description{
remove constraint-violating components from coefficient step delta
}
