% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{ConsIter}
\alias{ConsIter}
\title{ConsIter}
\usage{
ConsIter(
  A,
  grd,
  bas,
  C,
  g2,
  Ctol = 0.001,
  nsteps = 20,
  prn = FALSE,
  del_cons = 0.3,
  nm = FALSE,
  do_one = TRUE
)
}
\description{
Iterates coefficients to fulfill constraints.
Needed in Rosen Projection methods.
only SCM gradient is computed for constraint Jacobian
}
