% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{ConsJacobian}
\alias{ConsJacobian}
\title{ConsJacobian}
\usage{
ConsJacobian(g2, bas)
}
\arguments{
\item{g2}{energy E_SCM, containing gradients of constraints}

\item{bas}{basis functions like from M$bas}
}
\value{
Jacobian matrix made of row vectors from g2
}
\description{
constraint Jacobian for Rosen projection
}
