% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{ConstraintHessian}
\alias{ConstraintHessian}
\title{ConstraintHessian}
\usage{
ConstraintHessian(H, bas, Lambda, filter = ID)
}
\description{
constraint (=bordered) Hessian for CNM() Newton minimizer
}
