% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{Curv}
\alias{Curv}
\title{compute integrated curvature integ(k1+k2)}
\usage{
Curv(M)
}
\arguments{
\item{M}{MemRBC object}
}
\value{
total curvature
}
\description{
compute integrated curvature integ(k1+k2)
}
