% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{data_MemRBC}
\alias{data_MemRBC}
\title{data_MemRBC}
\usage{
data_MemRBC(name)
}
\arguments{
\item{name}{name of data (in quotes) to load}
}
\description{
load membrane data from package and set parameters
  - in contrast to data("name"), $Params from object are set to global params.
  - equivalent to : data("name");SetParams(name)
}
