% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{deltaX_norm}
\alias{deltaX_norm}
\title{deltaX_norm}
\usage{
deltaX_norm(C1, C2)
}
\arguments{
\item{C1, C2}{: coordinate objects}
}
\description{
Norm of coordinate change between two coordinate objects
}
