% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{DistBasedWX}
\alias{DistBasedWX}
\title{DistBasedWX}
\usage{
DistBasedWX(M)
}
\description{
computes spatial weights WX from 8 neares neighbors.
The weights may be used for subsampling or weighted fits.
Background: too dense gridpoints lead to underfitting in less denser regions.
}
