% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{-.MemRBC}
\alias{-.MemRBC}
\title{subtract coefficients}
\usage{
\method{-}{MemRBC}(m1, m2)
}
\arguments{
\item{m1, m2}{MemRBC objects for M1$A-M2$A}
}
\value{
result MemRBC object with coeffs difference
}
\description{
subtract coefficients
}
