% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{double_uv_ind}
\alias{double_uv_ind}
\title{double_uv_ind
locate double (u,v) indices, neccessary to filter for correct centre of mass
only double are allowed, i.e. tripples etc not detected}
\usage{
double_uv_ind(u, v, digits = 7)
}
\arguments{
\item{u, v}{vectors of spheiracl angular data, as from `MakeBasis_UV()`}

\item{digits}{(=7) precision of comparison checks}
}
\value{
indices vector of doubles (second entry taken), often storted to `M$bas$mask`
}
\description{
double_uv_ind
locate double (u,v) indices, neccessary to filter for correct centre of mass
only double are allowed, i.e. tripples etc not detected
}
