% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{E_FullModel_Penalty}
\alias{E_FullModel_Penalty}
\title{standardized model gradient with penalties}
\usage{
E_FullModel_Penalty(A, grd, bas, Ref)
}
\description{
standardized model gradient with penalties
}
