% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{E_FullModel_Penalty_AV}
\alias{E_FullModel_Penalty_AV}
\title{E_FullModel_Penalty_AV}
\usage{
E_FullModel_Penalty_AV(A, grd, bas, Ref)
}
\description{
Area-Volume-constraint energy computation
 works also without SEN (Ref=NULL)
}
