% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{E_FullModel_Penalty_AVC}
\alias{E_FullModel_Penalty_AVC}
\title{E_FullModel_Penalty_AVC}
\usage{
E_FullModel_Penalty_AVC(A, grd, bas, Ref)
}
\description{
Area-Volume-Curvature-constraint energy computation
}
