% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{E_SCM}
\alias{E_SCM}
\title{Bending energy}
\usage{
E_SCM(A, grd, bas, C, plt = FALSE, dbg = FALSE, clp = FALSE)
}
\arguments{
\item{A, grd, bas, C}{: coefficients, grid, basis and coordinates}

\item{plt}{(=FALSE) for plotting 3d}

\item{dbg}{(=FALSE) for debug output}

\item{clp}{(=FALSE) for plotting with a central clipping-plane}
}
\value{
list with "Wb" as bending energy and further data
}
\description{
compute bending energy
}
