% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{E_SCM_R}
\alias{E_SCM_R}
\title{Bending energy in R code}
\usage{
E_SCM_R(A, grd, bas, C, plt = FALSE, dbg = FALSE, clp = FALSE)
}
\arguments{
\item{A, grd, bas, C}{: coefficients, grid, basis and coordinates}

\item{plt}{(=FALSE) for plotting 3d}

\item{dbg}{(=FALSE) for debug messages}

\item{clp}{(=FALSE) for plotting with a central clipping-plane}
}
\description{
compute bending energy, returns list with "Wb" as bending energy
}
