% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{E_SEN}
\alias{E_SEN}
\title{E_SEN
energy of Shear-Elastic Network (SEN)}
\usage{
E_SEN(A, grd, bas, S, Ref)
}
\arguments{
\item{A, grd, bas, S, Ref}{Coefficients, grid, basis, SEN and reference objects}
}
\value{
list with alpha and beta as well as trace m of right Cauchy-Green
}
\description{
Compute the SEN energy
}
