% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Empty}
\alias{Empty}
\title{remove any App-cpmputational results from a MemRBC object.
The object is reduced to a minimum of data for further work/save.
Only grd,bas,Ref,A, Params and flag SEN are returned.}
\usage{
Empty(M)
}
\arguments{
\item{M}{MemRBC object}
}
\value{
data-reduced MemRBC object
}
\description{
remove any App-cpmputational results from a MemRBC object.
The object is reduced to a minimum of data for further work/save.
Only grd,bas,Ref,A, Params and flag SEN are returned.
}
