% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{Energy}
\alias{Energy}
\title{Energy}
\usage{
Energy(M)
}
\arguments{
\item{M}{The input membrane with initial data and reference}
}
\value{
vector of energies Wb (bending), Es (stress-shear), E (potetntial energy), Ekin (kinetic energy, optional, if a PNEM was run)
}
\description{
report  energy values for bending and SEN from a membrane object
}
\examples{
data("M1")
Energy(M1)
}
