% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{FillBasis_MemRBC}
\alias{FillBasis_MemRBC}
\title{FillBasis_MemRBC
insert full basis in M from LM stored in M$bas}
\usage{
FillBasis_MemRBC(M)
}
\arguments{
\item{M}{MemRBC object with
@export}
}
\description{
FillBasis_MemRBC
insert full basis in M from LM stored in M$bas
}
