% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{Filter_1_per_lm}
\alias{Filter_1_per_lm}
\title{filter for largest coeff. per lm row}
\usage{
Filter_1_per_lm(A, bas)
}
\arguments{
\item{A, bas}{coefficient matrix ans basis}
}
\value{
sparsified coefficient matrix
}
\description{
filter for largest coeff. per lm row
}
