% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{FitAlm_Tikhonov}
\alias{FitAlm_Tikhonov}
\title{FitAlm_Tikhonov
fit with regularization and weights:
filtering high frequencies in least squares
  lambda should be tested systematically by L curve discussion}
\usage{
FitAlm_Tikhonov(
  X,
  bas,
  lambda = 0,
  WX = rep(1, nrow(X)),
  keepIM = FALSE,
  newIM = FALSE
)
}
\arguments{
\item{X, bas}{spatial coords and basis}

\item{lambda}{(=0) regularization parameter}

\item{WX}{(=1) spatial weights, take e.g. sin(u)}

\item{keepIM}{(=FALSE) give back inversion matrix IM as attribute}

\item{newIM}{(=FALSE) if input basis has IM, it is preserved}
}
\value{
fitted coefficients object
}
\description{
FitAlm_Tikhonov
fit with regularization and weights:
filtering high frequencies in least squares
  lambda should be tested systematically by L curve discussion
}
