% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{FitDiscocyte_L5}
\alias{FitDiscocyte_L5}
\title{FitDiscocyte_L5}
\usage{
FitDiscocyte_L5(C0 = 0, V0 = 100)
}
\arguments{
\item{C0}{(=0) C0-value (spontaneous curvature) to use for initial minimization}

\item{V0}{(=100) target volume to use}
}
\value{
MemRBC object
}
\description{
Fit a discocyte from data, spectral order L=5
}
