% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{FitFast}
\alias{FitFast}
\title{FitFast}
\usage{
FitFast(bas, X)
}
\arguments{
\item{bas, X}{basis with IM, 3d-coordinates to fit}
}
\description{
fits coefficients to X from least squares solver matrix IM in bas
This is used for fitting e.g. displacements
like in pertA_complex.
}
