% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{FitProlate_L5}
\alias{FitProlate_L5}
\title{FitProlate_L5}
\usage{
FitProlate_L5(C0 = 2.562, V0 = 85.66, filter_strength = 10, no_minim = TRUE)
}
\arguments{
\item{C0}{spontaneous curvature for the model (default 2.562 for Undustick)}

\item{V0}{real volume (default 85.66 for Undustick)}
}
\description{
Fit coefficients from a stick with semi-spherical caps to initiate Undustick.
}
\examples{
# this sets also membrane parameters such that SEN is switched off
i.e. M.K_ADE=M.mu=M.Ka=0 # no SEN and ADE-term for Undustick
M.C0=2.562 # reference C0 value, see H0 in https://zenodo.org/records/13627757

FitProlate_L5() -> P5
P5
save_MemRBC(P5,"Undustick-prototype.rdat")
PlotLSeries(P5,3,2)
MMC(P5,100000)->P5mmc
PlotLSeries(P5,3,2)
SDRC(P5,1000,plot=TRUE,pltfreq=1)->P5_sdrc
P5_sdrc
}
