% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{FullHessian_Client}
\alias{FullHessian_Client}
\title{FullHessian_Client}
\usage{
FullHessian_Client(L)
}
\arguments{
\item{L}{List of settings to compute Hessian row from}
}
\value{
List of results, including gradient G for Hessian by finite difference for Hessian row index
}
\description{
Code exported for your inspection/modification.
Look into tmp_cluster.txt for its debug outputs from cluster nodes.
}
