% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{FullModelHessian}
\alias{FullModelHessian}
\title{FullModelHessian}
\usage{
FullModelHessian(A, grd, bas, Ref, del = 1e-06, Ctol = 0.001)
}
\arguments{
\item{A, grd, bas, Ref}{: standard objects of a MemRBC}

\item{del}{(=1e-6) finite difference delta for coefficients}

\item{Ctol}{(=0.001) constraint tolerance}
}
\value{
Full Hessian matrix, ie also constraints, but not bordered
}
\description{
computes the full model Hessian of the energy, as used in CNM.
It is based on finite differences of gradients
with a symmetrization; includes full Lagrangian.
}
