% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{FullModelHessian_Par}
\alias{FullModelHessian_Par}
\title{FullModelHessian_Par}
\usage{
FullModelHessian_Par(
  A,
  grd,
  bas,
  Ref,
  del = 5e-06,
  Mem_mc.cores = 4,
  timing = TRUE,
  startup = TRUE,
  stopdown = TRUE
)
}
\description{
parallel computed Hessian including constraint gradients; serial gradients but a list-parallel approach to assemble H
}
