% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAM.R
\name{GAM}
\alias{GAM}
\title{Genetic Algorithm Minizer}
\usage{
GAM(M, nsteps = 1, maxiter = 50)
}
\arguments{
\item{M}{The input membrane with initial data and reference}

\item{nsteps}{number of GA-cycles to be run}

\item{maxiter}{number of iterations in each GA cycle}
}
\value{
membrane with updates from GAM

LA: list of recorded coefficients A

GA: result of last GA cycle, needed for restart

A: last coefficients

history: history of App-calls that created the result
}
\description{
run GAM on a precomputed membrane object
}
\examples{
M <- MakeStandardRBC(L=5)
plot(M)
M <- GAM(M,nsteps=3)
plot(M)
M
}
