% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Gemini_cotan_Laplacian_I}
\alias{Gemini_cotan_Laplacian_I}
\title{Gemini_cotan_Laplacian_I
GEMINI created code:
from 3D-object mesh compute cotan-Laplacian}
\usage{
Gemini_cotan_Laplacian_I(mesh)
}
\arguments{
\item{mesh}{input 3d mesh object}
}
\value{
sparse cotan-Laplacian matrix
}
\description{
Gemini_cotan_Laplacian_I
GEMINI created code:
from 3D-object mesh compute cotan-Laplacian
}
