% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{GEMINI_cotan_Laplacian_II}
\alias{GEMINI_cotan_Laplacian_II}
\title{GEMINI_cotan_Laplacian_II
GEMINI created code:
from 3D-object mesh compute cotan-Laplacian}
\usage{
GEMINI_cotan_Laplacian_II(mesh)
}
\arguments{
\item{mesh}{input 3d mesh object}
}
\value{
sparse cotan-Laplacian matrix
}
\description{
GEMINI_cotan_Laplacian_II
GEMINI created code:
from 3D-object mesh compute cotan-Laplacian
}
