% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{GEMINI_disk_conformal_map}
\alias{GEMINI_disk_conformal_map}
\title{Map a 3D Mesh to a 2D Unit Circle (Disk Conformal Map)}
\usage{
GEMINI_disk_conformal_map(mesh, L = NULL, plt = FALSE, spherical = TRUE)
}
\arguments{
\item{mesh}{An Rvcg mesh3d object (must be an open mesh, not a closed sphere)}

\item{L}{The Cotangent Laplacian matrix. If NULL, it is computed.}

\item{spherical}{(=TRUE) for giving not uv on a disk but spherical coordinates (u,v)}
}
\value{
A list containing the 2D coordinates and the flattened mesh.
}
\description{
Map a 3D Mesh to a 2D Unit Circle (Disk Conformal Map)
}
