% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{GEMINI_Intersect_Mem_Mem}
\alias{GEMINI_Intersect_Mem_Mem}
\title{GEMINI_Intersect_Mem_Mem}
\usage{
GEMINI_Intersect_Mem_Mem(M1, M2 = M1)
}
\arguments{
\item{M1, M2}{MemRBC data}
}
\description{
translate membranes to 3D objects and check for intersection
}
