% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{get_data_ZENODO}
\alias{get_data_ZENODO}
\title{get_data_ZENODO}
\usage{
get_data_ZENODO(local = FALSE, folder = "data", L = NULL)
}
\arguments{
\item{local}{(=FALSE) to download in folder dir, no inst to packages data}

\item{folder}{(="data") name of local download folder (is kept on exit)}

\item{L}{(=NULL) specific files names for download (vector of character);Only published data can be requested. If NULL, all files are taken.}
}
\value{
character vector of all published files names.
}
\description{
get the additional data by downloading from ZENODO
and moving to data-folder.
}
