% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Grad_FullModel_Penalty}
\alias{Grad_FullModel_Penalty}
\title{unified gradient computation with penalties}
\usage{
Grad_FullModel_Penalty(A, grd, bas, Ref, S)
}
\description{
unified gradient computation with penalties
}
