% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Grad_SCM}
\alias{Grad_SCM}
\title{Grad_SCM}
\usage{
Grad_SCM(h2, grd, bas, C, int2d_m = .IntegM)
}
\arguments{
\item{h2}{: result from E_SCM()}

\item{grd, bas, C}{: coefficients, grid, basis and coordinates}

\item{int2d_m}{: integration function, e.g. .IntegM (hidden c++ R-function)}
}
\description{
compute bending energy gradient
}
