% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Grad_SCM_R}
\alias{Grad_SCM_R}
\title{Grad_SCM_R 
R version of SCM energy gradient, vectorizes first dimension (i = spatial)}
\usage{
Grad_SCM_R(Wb, grd, bas, C, int2d_m = int2d_matrix)
}
\arguments{
\item{Wb}{: result from E_SCM}

\item{grd, bas, C}{:  grid, basis and coordinates}

\item{int2d_m}{(=int2d_matrix) integration method}
}
\description{
compute bending energy gradient in R (slow)
}
