% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Grad_SEN}
\alias{Grad_SEN}
\title{Gradient of SEN energy}
\usage{
Grad_SEN(A, grd, bas, h2cur_grad, S, Ref, int2d_m = .IntegM)
}
\arguments{
\item{A, grd, bas}{Coefficients, grid, basis}

\item{h2cur_grad}{current SCM gradient (´Grad_SCM()´)}

\item{S}{: current ´SEN()´ result}

\item{Ref}{: static reference of SEN}

\item{int2d_m}{(.IntegM) matrix integrator function}
}
\value{
gradS for SEN energy gradient, gradAlpha, gradBeta, dm for gradients of stretches and shear.
}
\description{
Compute gradients of SEN energy
}
