% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{ID}
\alias{ID}
\title{ID (not used anymore)}
\usage{
ID(A, bas)
}
\arguments{
\item{A, bas}{coefficients, basis}
}
\value{
non-filtered coefficients
}
\description{
a template filter for modification of
coefficients.
Such a filter can be used in only a few apps, like MMC.
}
