% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{imag.delta.aligned}
\alias{imag.delta.aligned}
\title{imag.delta.aligned}
\usage{
imag.delta.aligned(A1, A2, grd2, bas2, shade2 = FALSE)
}
\arguments{
\item{A1, A2}{input coefficients}

\item{grd2, bas2}{grid and basis of A2}
}
\value{
obj1,obj2 aligned 3d-objects
}
\description{
svd alignment of shape coefficients into 3d-objects, see
https://www.cse.wustl.edu/~taoju/cse554/lectures/lect07_Alignment.pdf
and plotting of spatial difference vectors
}
\examples{
data(D5)
data(D5c5)
imag.delta.aligned(D5$A,D5c5$A,D5c5$grd,D5c5$bas)->L
rgl::open3d()
rgl::plot3d(L$obj1,alpha=0.6,col=1,aspect=FALSE)
rgl::plot3d(L$obj2,alpha=0.6,col=2,add=TRUE)
}
