% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{imag.obj.colorbar}
\alias{imag.obj.colorbar}
\title{imag.obj.colorbar
allow for limits to suppress outliers (color black)}
\usage{
imag.obj.colorbar(
  obj,
  f,
  limits = range(f),
  clr = FALSE,
  pal = heat.colors,
  width = 550,
  height = 480,
  par = FALSE,
  ...
)
}
\arguments{
\item{obj}{mesh3d object}

\item{f}{scalar quantity (spatial vector)}

\item{limits}{(=range(f)) limit displayed f}

\item{clr}{(=TRUE) clear3d() before plot}

\item{pal}{(=heat.colors) color palette}

\item{width, height}{window geometry}

\item{par}{(=FALSE) for par3d to take effect}

\item{...}{further parameters to shade3d()}
}
\value{
-
}
\description{
imag.obj.colorbar
allow for limits to suppress outliers (color black)
}
