% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{imag.obj.colorbar.simple}
\alias{imag.obj.colorbar.simple}
\title{plot on 3D object surface a field f as color code}
\usage{
imag.obj.colorbar.simple(obj, f, clr = TRUE, pal = rainbow, ...)
}
\arguments{
\item{obj}{mesh3d object}

\item{f}{scalar quantity (spatial vector)}

\item{clr}{(=TRUE) clear3d() before plot}

\item{pal}{(=rainbow) color palette}

\item{...}{further parameters to shade3d()}
}
\value{
-
}
\description{
plot on 3D object surface a field f as color code
}
