% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{imag.obj}
\alias{imag.obj}
\title{display f data on object obj}
\usage{
imag.obj(obj, f, pal = rainbow)
}
\arguments{
\item{obj}{mesh3d object}

\item{f}{scalar (spatial vecotr) to plot in color}

\item{pal}{(=rainbow) color palette to use}
}
\value{
-
}
\description{
display f data on object obj
}
