% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{imag}
\alias{imag}
\title{imag}
\usage{
imag(field, grd, nx = grd$nv, ny = grd$nu/2, ...)
}
\arguments{
\item{field}{scalar spatial field}

\item{grd}{the grid of (u,v) on which the scalar is defined}

\item{nx, ny}{dimension of output matrix plot}

\item{...}{additional parameters to `quilt.plot()`}
}
\value{
-
}
\description{
plot a scalar in 2D
}
