% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{int2d_matrix_cxx}
\alias{int2d_matrix_cxx}
\title{2d integration in C++}
\usage{
int2d_matrix_cxx(field_m, grd, bas)
}
\arguments{
\item{field_m}{matrix of spatial points to be integrated}

\item{grd, bas}{grid and basis containing relevant data for integration}
}
\value{
scalar value of integral of field_m
}
\description{
2d integration in C++
}
