% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{inv_sph}
\alias{inv_sph}
\title{inv_sph}
\usage{
inv_sph(X)
}
\arguments{
\item{X}{3d-coordinate}
}
\value{
uv vector of spherical angles c(u,v)
}
\description{
give angles  (u,v) from spherical or star-like shape 3D point X
}
