% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeStandardRBC.R
\name{invag_N}
\alias{invag_N}
\title{make invaginated shape MemRBC object
at the north pole by adding a gaussian displacement along Z}
\usage{
invag_N(X, width = 0.1, depth = 0.1)
}
\arguments{
\item{X}{spatial3d- coordinates}

\item{width}{(=0.1) width in micrometer}

\item{depth}{(=0.1) depth in micrometer}
}
\value{
modified spatial 3d coordinates
}
\description{
make invaginated shape MemRBC object
at the north pole by adding a gaussian displacement along Z
}
