% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeStandardRBC.R
\name{KleinB}
\alias{KleinB}
\title{Kleins bottle parameterization}
\usage{
KleinB(uv, b = 2, h = 6, plt = FALSE)
}
\arguments{
\item{uv}{n x 2 matrix of spherical angles (u,v) (u is internally scaled *2). uv can be created from `MakeBasis_UV()`}

\item{b}{(=2) width}

\item{h}{(=6) approx. height}

\item{plt}{(=FALSE) if 3d-plot is wanted}
}
\description{
Kleins bottle parameterization
}
