\name{L9_Stomatocyce_6_Eq}
\alias{L9_Stomatocyce_6_Eq}
\docType{data}
\title{
An MMC-equilibrated stomatocyte for C0=-6
}
\description{
This was created within MemRBC with FitStomatocyte_L(). See L9_Stomatocyce_6_Eq$history.
}
\usage{
data("L9_Stomatocyce_6_Eq")
}
\examples{
data("L9_Stomatocyce_6_Eq")
SetParams(L9_Stomatocyte_6_Eq)
plot(L9_Stomatocyce_6_Eq)
L9_Stomatocyce_6_Eq
}
\keyword{datasets}
