% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{last}
\alias{last}
\title{last -  return last n elements from v; v may be list}
\usage{
last(v, n = 1)
}
\arguments{
\item{v}{vector or list}

\item{n}{(=1) number of trailing elements to return}
}
\value{
elements from v
}
\description{
last -  return last n elements from v; v may be list
}
