% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{LM2A}
\alias{LM2A}
\title{annotates A with L-M-strings as rownames, + some attributes
sets rownames to coefficient matrix A}
\usage{
LM2A(A, bas)
}
\arguments{
\item{A, bas}{: coefficient matrix and matching basis, i,e, `bas$A` has the same sahape as `A`}
}
\description{
annotates A with L-M-strings as rownames, + some attributes
sets rownames to coefficient matrix A
}
