% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{load_MemRBC}
\alias{load_MemRBC}
\title{load_MemRBC}
\usage{
load_MemRBC(file = "MemRBC.rdat", unreduce_basis = TRUE)
}
\value{
the membrane object loaded from file with $Params to set globally
}
\description{
load membrane object from file and set global data from object
}
\examples{
M<-load_MemRBC("Mmmc.rdat")
M
}
