% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{loadA}
\alias{loadA}
\title{load a set of coeffs, make it compatible for use with the basis bas}
\usage{
loadA(file, bas)
}
\arguments{
\item{file, bas}{filename of coefficient file (binary .rdat) and basis of target object}
}
\value{
coefficient object (matrix)
}
\description{
load a set of coeffs, make it compatible for use with the basis bas
}
