% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{loadAlm}
\alias{loadAlm}
\title{loadAlm
loads coefficients and stores according to basis bas}
\usage{
loadAlm(file, bas)
}
\arguments{
\item{file}{filename of saved coefficients (binary .rdat)}

\item{bas}{basis, need not be compatible to saved coeffs}
}
\value{
coeffcicient matrix object from file
}
\description{
loadAlm
loads coefficients and stores according to basis bas
}
