% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Lowess_vcg_meanvbOBJ}
\alias{Lowess_vcg_meanvbOBJ}
\title{use polynomials of degree 12 to fit to lowess fits with other data.
usefull for plotting mean axisymmetric data along an axis}
\usage{
Lowess_vcg_meanvbOBJ(x, O)
}
\arguments{
\item{x}{coordinate to sample curvatures against (e.g. z-coords for Z-axisymmetric)}

\item{O}{3d mesh object}
}
\value{
List with x,y,k with x,y coordinates and k mean curvature, sorted by curvature
}
\description{
use polynomials of degree 12 to fit to lowess fits with other data.
usefull for plotting mean axisymmetric data along an axis
}
