% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{make_delta_normal_to_surface}
\alias{make_delta_normal_to_surface}
\title{make_delta_normal_to_surface}
\usage{
make_delta_normal_to_surface(delta, grd, bas, n)
}
\arguments{
\item{delta}{step in coefficients to project}

\item{grd, bas}{grid and basis}

\item{n}{current spatial normals, e.g. returned as $n in E_SCM}
}
\description{
normal motion filter by integrals
(contributed by C Woelper, 2024, University Bremen)
}
