% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeStandardRBC.R
\name{MakeDiscocyteRBC}
\alias{MakeDiscocyteRBC}
\title{MakeDiscocyteRBC}
\usage{
MakeDiscocyteRBC(L)
}
\arguments{
\item{L}{: spectral order of created object, L>5}
}
\description{
build from data M4 a higher spectral order discocyte object
}
