% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{MakeGrid_Fourier}
\alias{MakeGrid_Fourier}
\title{Fourier grid for genus 1 shapes
(not yet operative)}
\usage{
MakeGrid_Fourier(n = 30, r = 1, R = 2.5, check_plt = FALSE)
}
\arguments{
\item{n}{(=30) number of grid points in one direction}

\item{r}{(=1) for smaller radius of torus}

\item{R}{(=2.5) for larger radius of torus}

\item{check_plt}{(=FALSE) TRUE for verification 3d plot}
}
\value{
Grid object with (u,v) in $U, $V to compute basis from.
}
\description{
Fourier grid for genus 1 shapes
(not yet operative)
}
