% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{MakeGrid_GaussLegendre}
\alias{MakeGrid_GaussLegendre}
\title{MakeGrid_GaussLegendre
this has more dense points next to 0 meridian
 alternative with regular v-spacing is GaussLegendreSimpson}
\usage{
MakeGrid_GaussLegendre(
  n = 25,
  uv_fac = 1,
  comment = "spherical coordinates Gauss-Legendre grid, type GL",
  check_plt = FALSE
)
}
\arguments{
\item{n}{: number of points along each dimension}

\item{uv_fac}{(=1) : for experiemnts with half periods in v set uv_fac=0.5}

\item{comment}{(="spherical ...")}

\item{check_plt}{(=FALSE) for plotting}
}
\value{
2D Gauss-Legendre grid object
}
\description{
MakeGrid_GaussLegendre
this has more dense points next to 0 meridian
 alternative with regular v-spacing is GaussLegendreSimpson
}
