% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{MakeGrid_GaussLegendreSimpson}
\alias{MakeGrid_GaussLegendreSimpson}
\title{MakeGrid_GaussLegendreSimpson}
\usage{
MakeGrid_GaussLegendreSimpson(
  n = 20,
  ua = 0,
  ub = pi,
  va = 0,
  vb = 2 * pi,
  comment = "spherical coordinates Gauss-Legendre-Simpson grid, type GLS",
  check_plt = FALSE
)
}
\arguments{
\item{n}{(=20) : number of points along one dimension}

\item{ua, ub}{: usually 0,pi for u-interval}

\item{va, vb}{: usually 0,2*pi for v-interval}

\item{comment}{: give your own comment (as.character)}

\item{check_plt}{(=FALSE)}
}
\value{
list class MemGRD with grid parameters
}
\description{
compute an integration grid
}
