% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{MakeIM}
\alias{MakeIM}
\title{make inversion matrix for fast least squares}
\usage{
MakeIM(bas, lambda = 0, WX = rep(1, dim(bas$Ylm)[1]))
}
\arguments{
\item{bas}{Basis to compute inversion matrix from}

\item{lambda}{(=0) for additional diagonal term}

\item{WX}{(=1) for spatial weights}
}
\value{
modified basis, with result matrix in `$IM`
}
\description{
make inversion matrix for fast least squares
}
