% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeStandardRBC.R
\name{MakeKleinBottle}
\alias{MakeKleinBottle}
\title{Make a Klein bottle that is a bit open at the neck
the integratiön domain is not deltau...pi-deltau, i.e. caps of the spectral fit are suppressed}
\usage{
MakeKleinBottle(
  L = 17,
  n = L * 4,
  b = 2,
  h = 6,
  deltau = 0.15,
  plt = FALSE,
  SEN = FALSE
)
}
\arguments{
\item{L}{(=17) spectral order}

\item{detlau}{(=0.15) cap-cutting parameter}
}
\description{
Make a Klein bottle that is a bit open at the neck
the integratiön domain is not deltau...pi-deltau, i.e. caps of the spectral fit are suppressed
}
\examples{
M.C0=0;M.Ka=M.mu=0
M <- MakeKleinBottle(L=12,deltau=0.1)
#plot(M)
Quantities(M)
StoreParams(M)->M
M$bas$Target
M
update(M,"Obj") -> M
imag.obj.colorbar(M$grd$Obj,M$grd$v,pal=topo.colors,alpha=0.65)
}
