% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{MakeOBJ}
\alias{MakeOBJ}
\title{MakeOBJ
make a rgl-object from MemRBC data}
\usage{
MakeOBJ(A, grd, bas, C, col_wire = "white", col = "black", plot = FALSE)
}
\arguments{
\item{A, grd, bas, C}{coeffs, grid, basis and coordinates object}

\item{col_wire}{(="white") color of wireframe}

\item{col}{(="black") color of factes}

\item{plot}{(=FALSE) to show resulting shape in 3d}
}
\value{
mesh3d object, sometimes found in `C$Obj`
}
\description{
MakeOBJ
make a rgl-object from MemRBC data
}
