% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{MakeRef}
\alias{MakeRef}
\title{MakeRef}
\usage{
MakeRef(M, ARef)
}
\arguments{
\item{M}{MemRBC input object}

\item{ARef}{Reference shape coefficients for SEN}
}
\value{
MemRBC object with Ref and SEN set.
}
\description{
Build reference from input coefficients ARef and store in M.
If missing and demanded, M$Ref and M$SEN are computed here.
This helps plotting SEN colors like in PNEM and other Apps.
usually ARef comes from an slightly oblate shape of 90-95% of the target volume 100.
}
