% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{MakeSphere}
\alias{MakeSphere}
\title{MakeSphere}
\usage{
MakeSphere(grd, bas, r = 1)
}
\arguments{
\item{grd}{given grid}

\item{bas}{given basis}

\item{r}{(=1) for radius of output sphere}
}
\value{
coefficient matrix, derived from its prototype bas$A
}
\description{
Compute unit sphere for given grid and basis
}
