% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeStandardRBC.R
\name{MakeSpiculated}
\alias{MakeSpiculated}
\title{Create a spiculated cell from a sphere
via invag_N() with negative depth. The basic object started from is a L=16 standard sphere. The resulting shape is rescaled to obtain the target area specified in data(SS) (usually 140). You may also set a different volume target (see example code).
 WARNING: the reference is taken from unscaled sphere - may need updates!
NOTE: spatial points are weighted by sin(u) for the fit}
\usage{
MakeSpiculated(N = 42, w = 0.01, d = -0.35, r = 1, rz = 1)
}
\arguments{
\item{N}{(=20) (6,8,12,20,32,42) number of regularly arranged spikes}

\item{w}{(=0.01) width of spikes; 0.01 is good for N=42}

\item{d}{(=-0.35) negative elevation of spikes; 0.35 is good for N=42}

\item{r}{(=1) radius of initial sphere. Helps to control initial volume.}

\item{rz}{(=1) helps to scale z, rz<1 : oblate, rz>1 : prolate shape}
}
\description{
Create a spiculated cell from a sphere
via invag_N() with negative depth. The basic object started from is a L=16 standard sphere. The resulting shape is rescaled to obtain the target area specified in data(SS) (usually 140). You may also set a different volume target (see example code).
 WARNING: the reference is taken from unscaled sphere - may need updates!
NOTE: spatial points are weighted by sin(u) for the fit
}
\examples{
# example code for spiculated sphere ss
MakeSpiculated(N=12,w=0.06,d=-0.6,r=0.65) -> ss
ss
ss$bas$Target
#instead of tuning r,tz one may use ss$bas$Target["Volume"]=Volume(ss)

rgl::open3d()
plot(ss)
M.C0 <- 20
MMC(ss,100000,plt=TRUE,pltfreq=100,LAfreq=1000,C0=20) -> ss_mmc
}
