% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeStandardRBC.R
\name{MakeStandardRBC}
\alias{MakeStandardRBC}
\title{Make a standard RBC object}
\usage{
MakeStandardRBC(
  A0 = 140,
  V0 = 100,
  V0_Ref = 148,
  L = 9,
  C0 = -1,
  n = (L + 1) * 5 + 2,
  prn = FALSE,
  del_cons = 0.1,
  plt = FALSE
)
}
\arguments{
\item{A0}{to be constraint area}

\item{V0}{to be constraint volume}

\item{V0_Ref}{volume of reference shape (148 ~ v_0=0.95)}

\item{L}{spectral order l<=L}

\item{C0}{spontaneous curvature}

\item{n}{~sqrt(spatial grid points)}

\item{prn}{boolean, print infos on constraint iteration}

\item{del_cons}{(=0.1) constraint iterator step size}

\item{plt}{(=FALSE) for control plot}
}
\value{
membrane object with updates from MMC with:

LA: list of recorded coefficients A

A: last coefficients

bas: basis functions

grd: Gauss-Legendre-Simpson-spatial grid

Ref: Cauchy-Riemann tensor data for reference shape

ARef: coefficients of reference shape

history: history of App-calls that created the result
}
\description{
run MakeStandardRBC() to create a membrane object
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
MemRBC_env$M.C0 <- 0
# for fast experiments take a low-order L=5:
M <- MakeStandardRBC(L=5,prn=TRUE)
plot(M)
PSD(M,plt=TRUE)
M
\dontshow{\}) # examplesIf}
}
