% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeStandardRBC.R
\name{MakeStomatocyteRBC}
\alias{MakeStomatocyteRBC}
\title{make a stomatocyte from present data}
\usage{
MakeStomatocyteRBC(L = 5)
}
\arguments{
\item{L}{(=5) spectral order}
}
\description{
make a stomatocyte from present data
}
