% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeStandardRBC.R
\name{MakeTorus}
\alias{MakeTorus}
\title{Make torus membrane, i.e. genus 1
set SEN reference to initial shape, i.e., E_SEN=0}
\usage{
MakeTorus(L = 5, n = L * 6, R = 2.5, r = 1, plt = FALSE)
}
\arguments{
\item{L}{spectral order}

\item{n}{(=L*6) grid dimension}

\item{R}{(=2.5) greater radius of torus}

\item{r}{(=1) smaller radius of torus}

\item{plt}{(=FALSE) for control plots}
}
\value{
MemRBC object for a torus
}
\description{
Make torus membrane, i.e. genus 1
set SEN reference to initial shape, i.e., E_SEN=0
}
