% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{massmatrix}
\alias{massmatrix}
\title{massmatrix}
\usage{
massmatrix(M, rho = 1)
}
\arguments{
\item{M}{The input membrane with initial data and reference}

\item{rho}{area density, default 1}
}
\value{
massmatrix in spectral space
}
\description{
compute the mass matrix for PNEM
}
\examples{
data("M4")
mass <- massmatrix(m4,rho=2)
image(mass)
}
