% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{mat2vec}
\alias{mat2vec}
\title{mat2vec
helper routine for C code, making vector from matrix}
\usage{
mat2vec(m)
}
\arguments{
\item{m}{matrix}
}
\value{
vector from matrix data
}
\description{
mat2vec
helper routine for C code, making vector from matrix
}
