% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{matadd2vec}
\alias{matadd2vec}
\title{matadd2vec
helper routine for C code
gives back matrix sum as vector}
\usage{
matadd2vec(m1, m2)
}
\arguments{
\item{m1, m2}{matrices}
}
\value{
m1+m2 as vector
}
\description{
matadd2vec
helper routine for C code
gives back matrix sum as vector
}
