% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{matdiff2vec}
\alias{matdiff2vec}
\title{helper routine for C code
gives back matrix difference as vector}
\usage{
matdiff2vec(m1, m2)
}
\arguments{
\item{m1, m2}{matrices}
}
\value{
m1-m2 as vector
}
\description{
helper routine for C code
gives back matrix difference as vector
}
