% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Membrane_Curvatures}
\alias{Membrane_Curvatures}
\title{from SCM energy compute principal curvatures}
\usage{
Membrane_Curvatures(Wb, grd, plt.K = FALSE)
}
\arguments{
\item{Wb}{: result from E_SCM}

\item{grd}{:  grid for integration}

\item{plt.K}{(=FALSE) for plotting 2d}
}
\value{
list of curvature k1, k2, computed from K=L*N-M^2 (second fundamental)
}
\description{
compute principle curvatures k1,k2 from precomputed E_SCM
}
