% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Membrane_Eig}
\alias{Membrane_Eig}
\title{Membrane_Eig}
\usage{
Membrane_Eig(L, which = 1:4, kind = "SM")
}
\arguments{
\item{L}{mesh Laplacian}

\item{which}{(=1:4) : range of values to return, first 4 is default}

\item{kind}{(="SM") : e.g. for ordering starting from smallest magnitude}
}
\value{
matrix of demanded eigenvectors (columns); eigenvalues given as attribute "values"
}
\description{
get Eigensystem for mesh Laplacian (created from an rgl object -> Membrane_LaplacianOBJ())
}
