% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Membrane_Laplacian_cotan}
\alias{Membrane_Laplacian_cotan}
\title{Membrane_Laplacian_cotan
slow implementation in R}
\usage{
Membrane_Laplacian_cotan(x, M)
}
\arguments{
\item{x}{: mesh3d object}

\item{M}{: sparse matrix from Membrane_LaplacianOBJ, to fill with cotan values}
}
\value{
sparse cotan-Laplacian matrix
}
\description{
Membrane_Laplacian_cotan
slow implementation in R
}
