% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Membrane_LaplacianOBJ}
\alias{Membrane_LaplacianOBJ}
\title{Membrane_LaplacianOBJ
Laplacian matrix from graph}
\usage{
Membrane_LaplacianOBJ(X)
}
\arguments{
\item{X}{input 3d mesh object}
}
\value{
sparse Laplacian matrix
}
\description{
Membrane_LaplacianOBJ
Laplacian matrix from graph
}
