% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Membrane_LaplaciansOBJ}
\alias{Membrane_LaplaciansOBJ}
\title{Membrane_LaplaciansOBJ
from 3D-object X compute mesh Laplacian L and more}
\usage{
Membrane_LaplaciansOBJ(X)
}
\arguments{
\item{X}{input 3d mesh object}
}
\value{
List with L sparse Laplacian matrix, Ln normalized L, D =L with diagonal set to zero
}
\description{
Membrane_LaplaciansOBJ
from 3D-object X compute mesh Laplacian L and more
}
