% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{MemCols}
\alias{MemCols}
\title{Pallette color from data vector.}
\usage{
MemCols(data, pal = rainbow, n = 100)
}
\arguments{
\item{data}{data vector}

\item{pal}{(=rainbow) palette to use}

\item{n}{number of different colors}
}
\value{
vector of colors of length of data
}
\description{
Pallette color from data vector.
}
