% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MemRBC-package.R
\docType{package}
\name{MemRBC-package}
\alias{MemRBC}
\alias{MemRBC-package}
\title{MemRBC: Red Blood Cell shape modelling code laboratory}
\description{
MemRBC is an R-package for experiments in numerical modeling of the stomatocyte-discocyte-echinocyte-transformation of Red Blood Cell shape in a spectral model using spherical harmonics. It has a number of foundational implementations for energy, gradient and Hessian computation. The numerical algorithm applications for modeling tasks are offered as high-level functions ("Apps") for energy minimization as well as dynamics. Methods are Steepeds descend, Constrained Newton, Newton equation of motion, Genetic algorithm, Monte-Carlo, Augmented Lagrangian and variants for fixed curvature. An object-oriented approach is followed for these apps, in which metadata of previous modeling steps are kept in the MemRBC data. By this design, modularity of the modeling workflow, reproducibility and re-use is facilitated. Users are encouraged to modify the Apps or invent new ones.
}
\author{
\strong{Maintainer}: Stephan Frickenhaus \email{Stephan.Frickenhaus@awi.de} (ORCID=0000-0002-0356-9791)

Other contributors:
\itemize{
  \item Carl Woelper [contributor]
}

}
\keyword{internal}
