% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\docType{data}
\name{MemRBC_env}
\alias{MemRBC_env}
\title{package environment
to keep central parameters of models}
\format{
An object of class \code{environment} of length 22.
}
\usage{
MemRBC_env
}
\description{
package environment
to keep central parameters of models
}
\keyword{datasets}
