% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMC.R
\name{MMC}
\alias{MMC}
\title{MMC}
\usage{
MMC(
  M,
  nsteps = 1000,
  plt = TRUE,
  pltfreq = 10,
  prn = TRUE,
  LAfreq = 200,
  sd = 0.004,
  kT = 0.00411,
  kTfac = 1,
  kTfreq = 100,
  pertA = pertA_Unif,
  record_dA = FALSE,
  timing = FALSE,
  C0 = M$C0,
  ...
)
}
\arguments{
\item{M}{The input membrane with initial data and reference}

\item{nsteps}{number of MMC steps to be run}

\item{plt}{(=FALSE) for plotting}

\item{prn}{(=TRUE) for extensive printing}

\item{LAfreq}{(=200) for storing shape coefficients every LAfreq accepted steps}

\item{sd}{(=0.004) standard deviation parameter to control steps}

\item{kT}{(=0.00411) Boltzmann-constant * temperature energy scale}

\item{kTfac}{(=1.0) cooling factor (0<kTfac<1) or heating factor (kTfac>1)}

\item{plfreq}{(=5) for plotting shape every plotfreq number of accepted steps}

\item{nm}{(=FALSE) normal motion filter}

\item{filter}{(=ID) no other filter implemented, do it yourself}
}
\value{
membrane object with updates from MMC with data:

LA: list of recorded coefficients A

A: final coefficients

kT:  value of kT on exit, important for restarting

MMCacceptanceRate: acceptance rate of the call

MMCiter: total MMC steps, incl. MMC from previous calls

history: history of App-calls that created the result
}
\description{
Run Metropolis Montecarlo MMC on an existing membrane object under penalty constraints for volume and area.
 Constraints are set in M$bas$Target for penalization with strength M.rho.
}
\examples{
M <- MakeStandardRBC(L=5)
plot(M)
#  annealing simulation (decrease kT by kTfac every Ktfreq accepted steps)
M1 <- MMC(M,nsteps=100000, kT=0.00411, kTfac=0.99, kTfreq=100, C0=-2)
plot(M1)
M1
}
