% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Movie}
\alias{Movie}
\title{Movie}
\usage{
Movie(
  M,
  x = 400,
  y = 400,
  from = 1,
  to = length(M$LA),
  sleep = 0.025,
  phi = pi/100,
  phi1 = phi/2,
  ...
)
}
\arguments{
\item{M}{MemRBC object}

\item{x, y}{window dimensions in pixels}

\item{from, to}{frames to show}

\item{sleep}{(=0.025) pause between images}

\item{phi}{(=pi/100) rotation angle per image}

\item{phi1}{(=phi/2) second rotation angle per image}

\item{...}{parameters to plot(M)}
}
\value{
-
}
\description{
show shapes along recorded coefficients in M$LA
}
