% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMC.R
\name{NNuv}
\alias{NNuv}
\title{NNuv}
\usage{
NNuv(uv, n = 13)
}
\arguments{
\item{uv}{(n x 2) matrix of (u,v) spherical angles}

\item{n}{(=13) number of nearest neighbours to compute}
}
\value{
result from RANN::nn2: $nn.idx and $nn.dists
}
\description{
compute nearest neighbors for pertA_complex
 store in global M.nn
 ATTENTION: returned distances $nn.dists are weighted sin(u), because in pole vicinity points get denser
}
