% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Obj2ObjQ}
\alias{Obj2ObjQ}
\title{Obj2ObjQ}
\usage{
Obj2ObjQ(O, grd)
}
\arguments{
\item{O, grd}{3d mesh object O and grid grd}
}
\value{
3d object with quads
}
\description{
makes a 3d-object of quads from a 3d-object of triangles (requires regular grid grd)
}
