% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Obj2X}
\alias{Obj2X}
\title{helper routines for 3d graphics objects}
\usage{
Obj2X(O)
}
\arguments{
\item{O}{mesh3d object}
}
\value{
3d-coordinates
}
\description{
helper routines for 3d graphics objects
}
