% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Obj2X_centre}
\alias{Obj2X_centre}
\title{make coordinates from centered object}
\usage{
Obj2X_centre(O)
}
\arguments{
\item{O}{mesh3d object}
}
\value{
centered 3d-coordinates
}
\description{
make coordinates from centered object
}
