% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Obj_Obj_Intersect}
\alias{Obj_Obj_Intersect}
\title{KD-Tree accelerated triangle-triangle-intersection search}
\usage{
Obj_Obj_Intersect(S1, S2 = S1)
}
\arguments{
\item{S1, S2}{mesh3d objects to check for intersection; without S2 self intersection is checked}
}
\description{
KD-Tree accelerated triangle-triangle-intersection search
}
