% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMC.R
\name{pertA_complex}
\alias{pertA_complex}
\title{pertA_complex}
\usage{
pertA_complex(A, bas, sd, n, nn = 12)
}
\arguments{
\item{A, bas}{coefficients and basis}

\item{sd}{standard deviation, not needed here}

\item{n}{normal vectors}

\item{nn}{(=12) number of nearest neighbours to construct displacements from}
}
\value{
modified coefficients A+dA
}
\description{
Perturb with gaussian along spatial normals.
 BEFORE USE:
a) Normals of neighbours require to fill the global M.nn=Nuv(bas$uv)
b) the basis must contain a matrix IM for least squares inversion in FitFast();
Use bas<-MakeIM(bas,WX=sin(grd$U)) to create IM.
}
