% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMC.R
\name{pertA_Gauss}
\alias{pertA_Gauss}
\title{perturb coeffs}
\usage{
pertA_Gauss(A, bas, sd, n, ...)
}
\arguments{
\item{A, bas}{coeffs and basis}

\item{sd}{standard dev. of gaussian}

\item{n}{not used}

\item{...}{not used}

\item{flt}{(=FALSE) to filter for only 3 entries in l=1}
}
\description{
perturb coeffs
}
