% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMC.R
\name{pertA_rndX}
\alias{pertA_rndX}
\title{pertA_rndX}
\usage{
pertA_rndX(A, bas, sd, n, sample_fraction = 0.65)
}
\description{
Perturb coeffs from spatial coordinates.
 BEFORE USE:
the basis mus contain a matrix IM for least squares inversion in FitFast();
Use bas<-MakeIM(bas,WX=sin(grd$U)) to create IM.
}
