% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMC.R
\name{pertA_Unif}
\alias{pertA_Unif}
\title{perturb coeffs A by dA from uniform distribution, as needed in GAM and MMC
Perturbation dA is scaled down by `sqrt(bas$G.tk)`}
\usage{
pertA_Unif(A, bas, sd, n, ...)
}
\arguments{
\item{A, bas}{coefficients and basis}

\item{sd}{width of perturbation}

\item{n}{normal vectors, e.g.from `E_SCM()`}

\item{...}{not used}
}
\value{
changed coeffs A+dA
}
\description{
perturb coeffs A by dA from uniform distribution, as needed in GAM and MMC
Perturbation dA is scaled down by `sqrt(bas$G.tk)`
}
