% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{plot3a}
\alias{plot3a}
\title{plot3a}
\usage{
plot3a(
  X,
  grd,
  pnts = FALSE,
  clip = FALSE,
  col = "black",
  cont = TRUE,
  cont.grid = FALSE,
  fill = TRUE,
  fn = "z",
  fn_data = "z",
  ...
)
}
\arguments{
\item{X}{Coordiates, e.g. in C$X from updateX()}

\item{grd}{grid with a basic rgl-object grd$Obj}

\item{pnts}{(=FALSE) for grid points to plot}

\item{clip}{(=FALSE) for clipplane (0.5,0.5,0)}

\item{col}{(="black")}

\item{cont}{(=TRUE) for contours of X,Y,Z levels}

\item{cont.grid}{(=FALSE) for u,v grid lines}

\item{fill}{(=TRUE) to fill contours}

\item{fn}{(="z") parameter of filled contours}

\item{fn_data}{(="z") for data as filled contour parameter}

\item{...}{parameters to filledContour3d()}
}
\description{
plot coordinates C$X (after updateX)
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
data("M4",package = "MemRBC")  # take required data from M5
plot3a(updateX(M4$A,M4$grd,M4$bas)$X,M4$grd)
\dontshow{\}) # examplesIf}
}
