% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{plot3b}
\alias{plot3b}
\title{plot3b}
\usage{
plot3b(X, grd, col = "white", specular = "black", wire = TRUE, ...)
}
\arguments{
\item{X, grd}{: 3d coordinates and 2d grid object}

\item{col, specular}{: (="white", ="black") graphics colors}

\item{wire}{(=TRUE) to plot wireframe on surface}

\item{...}{: further plotting options, e.g. alpha=0.5 for semi-transparency}
}
\description{
plot a 3d shape from coordinates C$X
}
