% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{plot3q}
\alias{plot3q}
\title{plot3q}
\usage{
plot3q(X, grd, col = "black", ...)
}
\arguments{
\item{X}{Coordiates, e.g. in C$X from updateX()}

\item{grd}{grid with a basic rgl-object grd$Obj}

\item{col}{(="black") wireframe color}

\item{...}{further parameters to plot3b()}
}
\value{
-
}
\description{
plot coordinates C$X (after updateX) with wireframe of quadrilaterals
}
\examples{
data("M4",package = "MemRBC")  # take required data from M4
plot3q(updateX(M4$A,M4$grd,M4$bas)$X,M4$grd)
}
