% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{plot3qs}
\alias{plot3qs}
\title{plot3qs}
\usage{
plot3qs(X, grd, s, alpha = 1, specular = "black", ...)
}
\arguments{
\item{X}{Coordiates, e.g. in C$X from updateX()}

\item{grd}{grid with a basic rgl-object grd$Obj}

\item{s}{scalar to plot as color code on shape}
}
\description{
plot with coordinates C$X (after updateX) and scalar as color code
}
\examples{
data("M1"); SetParams(M1)
 # take required data from M1
 update(M1,"dA")->M1
 #plot area sizes as color code
plot3qs(updateX(M1$A,M1$grd,M1$bas,M1$)$X,M1$grd,M1$dA)

}
