% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{plot_A}
\alias{plot_A}
\title{plot coeffs from coeffs object
- instead you may use plotA(M) on a MemRBC object M}
\usage{
plot_A(A, ...)
}
\arguments{
\item{A}{coeffs to plot}

\item{...}{further params to matplot()}
}
\value{
-
}
\description{
plot coeffs from coeffs object
- instead you may use plotA(M) on a MemRBC object M
}
