% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{PlotA}
\alias{PlotA}
\title{Plot coefficients by matplot(). 
Colors black, red, green for X, Y, Z coefficients.}
\usage{
PlotA(M, bar = TRUE, scale_up = TRUE, ...)
}
\arguments{
\item{M}{membrane to plot spectral coefficients M$A from}

\item{bar}{(=TRUE) for L-levels to be color-plotted (bar at y=0)}

\item{scale_up}{(=TRUE) plot A scaled-up by sqrt(bas$G.tk)}

\item{...}{parameters handed to matplot}
}
\value{
-
}
\description{
Plot coefficients by matplot(). 
Colors black, red, green for X, Y, Z coefficients.
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
data("M4",package = "MemRBC",envir=environment());
SetParams(M4)
PlotA(M4,scale_up=FALSE)
\dontshow{\}) # examplesIf}
}
