% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{plotA_l}
\alias{plotA_l}
\title{plotA_l
plot coeffs and l-value ranges as colored bars on the x-axis}
\usage{
plotA_l(
  A,
  bas,
  bar = FALSE,
  xlab = ifelse(all(diff(bas$Lset) == 1), expression((l + 1)^2 - l + m), "k"),
  scale_up = FALSE,
  ...
)
}
\arguments{
\item{A, bas}{: standard coefficient amd basis functions objects}

\item{bar}{(=FALSE)}

\item{xlab}{(=...) for standard formula index(l,m)}

\item{scale_up}{(=FALSE) to scale up by sqrt(bas$G.Tk)}

\item{...}{parameters for matplot, e.g. ylab}
}
\value{
-
}
\description{
plotA_l
plot coeffs and l-value ranges as colored bars on the x-axis
}
