% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{PlotDiff}
\alias{PlotDiff}
\title{PlotDiff}
\usage{
PlotDiff(M1, M2, On3d = FALSE, ...)
}
\arguments{
\item{M1, M2}{membrane objects; displayed delta is M2$A-M1$A}

\item{On3d}{(=FALSE) to plot on M1 surface the |dX| in color code}

\item{...}{graphics parameters to wire3d() or imag.obj.colorbar()}
}
\value{
named vector of changes of Quantities()
}
\description{
plot the spatial difference vectors as wireframe.
A quite simmilar plot can be obtained by plot(M2-M1).
}
\examples{
data(ss42denovo)
data(ss42denovo_pnem)
dQ <- PlotDiff(ss42denovo_pnem,ss42denovo,col="red")
rgl::open3d()
plot(ss42denovo_pnem-ss42denovo,col="green")
}
