% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{PlotLSeries}
\alias{PlotLSeries}
\title{PlotLSeries}
\usage{
PlotLSeries(M, nr = 4, nc = 3, ...)
}
\arguments{
\item{M}{membrane object to plot truncation series from}

\item{nr, nc}{(=4,=3) number of rows and columns on screen}
}
\description{
plot series of shapes from a single object for truncated to lower spectral orders
}
\examples{
data("M4")
M4
PlotLSeries(M4,2,3)
}
