% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{PlotPaths}
\alias{PlotPaths}
\title{PlotPaths}
\usage{
PlotPaths(O, P, LVPath = TRUE)
}
\arguments{
\item{O}{3d-mesh object}

\item{P}{Path computed on O, like from `NorthSouth()`}

\item{LVPath}{(=TRUE) to plot P$LVPath}
}
\value{
-
}
\description{
draw some paths P on object O
}
