% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{PlotPSD}
\alias{PlotPSD}
\title{PlotPSD}
\usage{
PlotPSD(M, lastE = length(M$E_PSD), lastC = lastE)
}
\arguments{
\item{M}{membrane object to plot PSD recorded energies from}

\item{lastE}{(length M$E_PSD) number of trailing energy points to plot}

\item{lastC}{(=lastE) number of trailing curvature points to plot}
}
\description{
plot recorded energies from PSD runs
}
\examples{
data("M4")
PSD(M4,1000) -> M
PSDC(M, curv=Curv(M), 1000) -> M1
PlotPSD(M1, lastE=750)
}
