% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{PlotRef}
\alias{PlotRef}
\title{PlotRef}
\usage{
PlotRef(M, ARef = M$Ref$ARef)
}
\arguments{
\item{M}{membrane object to plot SEN reference shape from}

\item{Aref}{optional coefficients, eg if the objects ARef is not working}
}
\description{
plot SEN reference shape from membrane object
}
\examples{
data("M4")
PlotRef(M4)
}
