% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{PlotSample}
\alias{PlotSample}
\title{PlotSample}
\usage{
PlotSample(M, last = dim(M$Sample)[1], title = "MMC sample plot", ...)
}
\arguments{
\item{M}{membrane object with $Sample keeping the MMC recorded data}

\item{title}{title of the plot, placed in separate box}
}
\value{
-
}
\description{
plot the curvature-energy-data sampled from MMC.
}
\examples{
data("M4")
MMC(M4,1000)->M4mmc
PlotSample(M4mmc)
}
