% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDRC.R
\name{PlotSDRC}
\alias{PlotSDRC}
\title{plot recorded data from SDRC}
\usage{
PlotSDRC(
  M,
  what = c("E", "C"),
  from = 1,
  to = dim(M$SDRC_Sample)[1],
  nx = 2,
  ny = 1,
  ...
)
}
\arguments{
\item{M}{MemRBC object containing SDRC data in $SDRC_Sample}

\item{what}{(=`c("E","C")`) besides energy and curvature, volume and area can be requested, see names(M$SDRC_Sample) with CN = Constraint Norm, CI = Constraint Iteration count, ID = run-ID}

\item{from, to}{specifies the step number range to plot}

\item{nx, ny}{number of plots on screen by `par(mfrow(ny,nx))`}

\item{...}{additional plot parameters, like `type="l"`}
}
\value{
-
}
\description{
plot recorded data from SDRC
}
