% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{PlotStabGallery}
\alias{PlotStabGallery}
\title{Plot a stability analysis visually on a 3x3 canvas of 3d-plots}
\usage{
PlotStabGallery(M, which_n = 1:9, sharedMouse = TRUE, wire = FALSE, ...)
}
\arguments{
\item{M}{The input membrane with initial data and reference}

\item{which_n}{vector of Ids of eigenvalues, 1:9 is default for lowest nine}

\item{sharedMouse}{(boolean) if FALSE, each subplot can be controlled by mouse individually}

\item{plt_scale}{scale of eigenvectors}

\item{col1}{color of unperturbed shapes}

\item{col2}{color of scaled eigenvector perturbed shapes}
}
\value{
none
}
\description{
run PlotStabGallery on a membrane object with a Stab in it from MemStab()
}
\examples{
M <- MakeStandardRBC(L=5)
plot(M)
M <- PSD(M,nsteps=100,del=1e-6,)
plot(M)
MemStab(M)->M
PlotStabGallery(M)
}
