% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{+.MemRBC}
\alias{+.MemRBC}
\title{add coefficients}
\usage{
\method{+}{MemRBC}(m1, m2)
}
\arguments{
\item{m1, m2}{MemRBC objects to add}
}
\value{
result MemRBC object
}
\description{
add coefficients
}
