% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{print.MemA}
\alias{print.MemA}
\title{print some coefficient metadata class MemA}
\usage{
\method{print}{MemA}(x, ...)
}
\arguments{
\item{x}{coefficient matrix object class "MemA"}

\item{...}{not needed}
}
\value{
-
}
\description{
print some coefficient metadata class MemA
}
