% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{print.MemBas}
\alias{print.MemBas}
\title{print some basis metadata class MemBas}
\usage{
\method{print}{MemBas}(x, ...)
}
\arguments{
\item{x}{basis object of class "MemBas"}

\item{...}{not needed}
}
\value{
-
}
\description{
print some basis metadata class MemBas
}
