% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{print.MemC}
\alias{print.MemC}
\title{print some coordinate object metadata class MemC}
\usage{
\method{print}{MemC}(x, ...)
}
\arguments{
\item{x}{coordinate object of class "MemC", like from `updateX()`}

\item{...}{not needed}
}
\value{
-
}
\description{
print some coordinate object metadata class MemC
}
