% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{print.MemC_X}
\alias{print.MemC_X}
\title{print some coordinate object metadata, where only X is set (class MemC_X)}
\usage{
\method{print}{MemC_X}(x, ...)
}
\arguments{
\item{x}{grid object of class "MemC_X"}

\item{...}{not needed}
}
\value{
-
}
\description{
print some coordinate object metadata, where only X is set (class MemC_X)
}
