% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{print.MemESCM}
\alias{print.MemESCM}
\title{print some SCM Energy metadata class MemESCM}
\usage{
\method{print}{MemESCM}(x, ...)
}
\arguments{
\item{x}{Energy object class "MemESCM"}

\item{...}{not needed}
}
\value{
-
}
\description{
print some SCM Energy metadata class MemESCM
}
