% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{print.MemGrd}
\alias{print.MemGrd}
\title{print some grid metadata}
\usage{
\method{print}{MemGrd}(x, ...)
}
\arguments{
\item{x}{grid object}

\item{...}{not needed}
}
\value{
-
}
\description{
print some grid metadata
}
