% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{print.MemRBC}
\alias{print.MemRBC}
\title{print}
\usage{
\method{print}{MemRBC}(x, ...)
}
\arguments{
\item{x}{Membrane to print data from}

\item{...}{not used}
}
\description{
print metadata from membrane object class MemRBC
}
