% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{print.MemRef}
\alias{print.MemRef}
\title{print some Reference metadata class MemRef}
\usage{
\method{print}{MemRef}(x, ...)
}
\arguments{
\item{x}{grid object of class MemRef as create e.g. by `MakeRef()`}

\item{...}{not needed}
}
\value{
-
}
\description{
print some Reference metadata class MemRef
}
