% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSDC.R
\name{PSDC}
\alias{PSDC}
\title{Penalized Steepest Descend Curvature}
\usage{
PSDC(
  M,
  curv = Curv(M),
  nsteps = 100,
  del = 1e-06,
  plt = FALSE,
  pltfreq = 10,
  LAfreq = 100,
  ncores = 5
)
}
\arguments{
\item{M}{The input membrane with initial data and reference}

\item{curv}{(=Curv(M))target curvature}

\item{nsteps}{(=100) number of PSDC steps to be performed}

\item{del}{(=1e-6) step-size}

\item{plt}{(=FALSE) for plotting}

\item{pltfreq}{(=10) for plotting every pltfreq step}

\item{LAfreq}{(=100) storage frequency into list of coefficients $LA}

\item{ncores}{(=4) parallel threads in SCM energy and gradient}
}
\value{
membrane object with updates from MMC with data:

LA: list of recorded coefficients A

A: last coefficients, with further data in it as attributes (see example)

C: last curvature

E: last total energy

E_PSD: all recorded total energies, including PSDC steps

PSDiter; total PSD steps, incl. previous calls

history: history of App-calls that created the result
}
\description{
run PSDC on an existing membrane object to reduce energy by gradient steps.
Here, curvature is an additional constraint (by penalty).
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
data("M4",package = "MemRBC")
M <- PSDC(M4,nsteps=10000,del=1e-6,)
plot(M)
M
attributes(last(M$LA))
attr(M$A,"method")
\dontshow{\}) # examplesIf}
}
