% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{radial_uv}
\alias{radial_uv}
\title{radial_uv
give angles u,v from a starlike 3d-object, to be centred}
\usage{
radial_uv(starlike_obj)
}
\arguments{
\item{starlike_obj}{a mesh3d object where all points can be reached uniquely by a straight line from center of mass}
}
\value{
uv (u,v) matrix as derived from coordinates of starlike_obj
}
\description{
radial_uv
give angles u,v from a starlike 3d-object, to be centred
}
