% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Reduce_Basis}
\alias{Reduce_Basis}
\title{Reduce_basis
remove unwanted Ylm
 e.g. for point symmetry, remove all even l}
\usage{
Reduce_Basis(bas, A, w)
}
\description{
Reduce_basis
remove unwanted Ylm
 e.g. for point symmetry, remove all even l
}
\examples{
# remove even l from basis of undustick model U17R
data(U17R)
SetParams(U17R) # sets model constants
Volume(U17R)
Area(U17R)
U17R$bas$Target
red<-Reduce_Basis(U17R$bas,U17R$A,which(U17R$bas$l\%\%2==0))
U17R$A <- red$A # feed back to MemRBC object
U17R$bas <- red$bas
plot(U17R) # does it still work?
Volume(U17R)
red$excluded_Ylm # which l,m where excluded from basis
}
