% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{ReduceM1}
\alias{ReduceM1}
\title{ReduceM1}
\usage{
ReduceM1(M)
}
\arguments{
\item{M}{membrane object to change}
}
\value{
modified M with reduced basis, but $Ref unchanged.
}
\description{
Reduce basis in M to leading cos v, sin v terms, i.e. |m|<2 in Y_lm.
Requires update(M,"Ref"), if SEN is used (i.e. M.Ka and/or M.mu not zero.
It is recommended to save a reduced model with save_MemRBC(...,reduce_basis=FALSE),
to keep the basis in reduced form in the file.
For loading, you must then use load_MemRBC(..., unreduce_basis=FALSE) to not overwrite the reduced basis.
 
The reduction on X and Y coeficients makes a Axisymmetric parameterization.
 Take care: The reduction is also on Z coeffs, 
 which drastically reduces the representable shape space.
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
#download data with 
if (interactive()) {
get_data_ZENODO(L="U17R.rda",local=TRUE)
load_MemRBC("data/U17R.rda")->U17R;
SetParams(U17R)
ReduceM1(U17R) -> L10fast
plot(L10fast)
 MMC(L10fast, 20000,C0=2.562)->M10fast_mmc
 PlotSample(L10fast_mmc,last=20000)
}
\dontshow{\}) # examplesIf}
}
