% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Ref4CauchyGreen}
\alias{Ref4CauchyGreen}
\title{Ref4CauchyGreen
compute lists of reference tensors for fast SEN computation}
\usage{
Ref4CauchyGreen(A, grd, bas)
}
\arguments{
\item{A, grd, bas}{: coefficients, grid and basis as from MemRBC object}
}
\value{
Reference object containing pieces of Cauchy-Green-tensors
}
\description{
Ref4CauchyGreen
compute lists of reference tensors for fast SEN computation
}
