% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Rosen_filter_delta}
\alias{Rosen_filter_delta}
\title{Rosen_filter_delta}
\usage{
Rosen_filter_delta(delta, grd, bas, H2, nm = TRUE)
}
\arguments{
\item{delta}{the step vector to filter for constraint violation}

\item{grd, bas}{grid and basis, like from `M$grd`, `M$bas`}

\item{H2}{current `E_SCM()`}

\item{nm}{(=TRUE) for normal motion filter}
}
\value{
projected delta as n x 3 matrix
}
\description{
remove constraint-violating components from coefficient step delta
the delta is always scaled down `1/(1+MemRBC_env$ll.a*sqrt(bas$G.tk)`
}
