% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{rotateA}
\alias{rotateA}
\title{rotateA
rotate coefficients by rotating coordinates
 gives back a rotation error (new from fit vs. rotated coord)
 rotation order is by X,Y,Z-axis
 a much more precise rotation would apply Wigner D-matrices}
\usage{
rotateA(A, bas, grd, px = pi/2, py = -pi/2, pz = pi/2, plt = FALSE)
}
\arguments{
\item{A, bas, grd}{: coeffs, basis, grid}

\item{px, py, pz}{. rotation angles}

\item{plt}{(=FALSE) for 3d platting resulting shape}
}
\value{
List with $A coefficient matrix, $C coordinate object (C$X for points)
}
\description{
rotateA
rotate coefficients by rotating coordinates
 gives back a rotation error (new from fit vs. rotated coord)
 rotation order is by X,Y,Z-axis
 a much more precise rotation would apply Wigner D-matrices
}
