% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{rotateX}
\alias{rotateX}
\title{rotateX
rotation first around x by px, then y then z-axis}
\usage{
rotateX(X, px = pi, py = 0, pz = 0, transpose = FALSE)
}
\arguments{
\item{X}{input coordinates matrix (n x 3)}

\item{px, py, pz}{rotation angles}

\item{transpose}{(=FALSE) if transposed matrices are wanted (for reverse rotation)}
}
\value{
rotated X
}
\description{
rotateX
rotation first around x by px, then y then z-axis
}
