% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotUV.R
\name{rotUV}
\alias{rotUV}
\title{rotUV}
\usage{
rotUV(Min, du, dv, plt = FALSE, transpose = FALSE)
}
\arguments{
\item{Min}{membrane object to rotate}

\item{du}{angles of rotation relative to Z-axis}

\item{dv}{angle of rotation around Z-axis}

\item{plt}{(=FALSE) TRUE for 3d plotting}

\item{transpose}{(=FALSE) for backward rotation, not verified}
}
\value{
membrane MemRBC object with coefficients after rotation, but original SEN reference
}
\description{
rotate membrane coordinates and re-compute coefficients
without modifying the SEN reference.
rotUV may help to move poles to Z-axis or for more homogeneous distribution of spatial points.
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
data("M4",package = "MemRBC")
plot(M4)
rotUV(M4,pi/2,3*pi/4)->M
plot(M)
\dontshow{\}) # examplesIf}
}
