% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{rvcg_kdtree_candidates}
\alias{rvcg_kdtree_candidates}
\title{KD-Tree accelerated candidate search}
\usage{
rvcg_kdtree_candidates(meshA, meshB)
}
\arguments{
\item{meshA, meshB}{mesh3d objects}
}
\description{
KD-Tree accelerated candidate search
}
