% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{save_MemRBC}
\alias{save_MemRBC}
\title{save_MemRBC}
\usage{
save_MemRBC(M, file = "MemRBC.rdat", reduce_basis = TRUE)
}
\arguments{
\item{M}{The input membrane to be saved}
}
\description{
save membrane object M to file
erases data from basis M$bas (recomputed when loaded back)
}
\examples{
data("M4")
MMC(M4,10)->M
save_MemRBC(M,"Mmmc.rdat")
}
