% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{saveA}
\alias{saveA}
\title{saveA
save a set of coefficients A to file}
\usage{
saveA(A, file)
}
\arguments{
\item{A, file}{coefficient object and filename}
}
\description{
saveA
save a set of coefficients A to file
}
