% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDRC.R
\name{SDRC}
\alias{SDRC}
\title{SDRC}
\usage{
SDRC(
  M,
  nsteps = 100,
  del_min = 1e-07,
  del_cons = 0.001,
  Gtol = 1e-04,
  max_iter = 10,
  cons_tol = 0.001,
  plt = FALSE,
  pltfreq = 10,
  LAfreq = 25,
  filter_strength = 0,
  prn_ci = FALSE
)
}
\arguments{
\item{M}{The input membrane with initial data and reference}

\item{nsteps}{number of SD-steps to be run (with varying number of constraint iteration steps CI in between)}

\item{del_min}{(=1e-7) step-size of SD}

\item{del_cons}{(=1e-3) step-size of fix point iteration}

\item{Gtol}{(=1e-4) stop if gradient norm is below Gtol}

\item{max_iter}{(=10) maximum number of constraint iterations}

\item{cons_tol}{(=1e-3) constraint iterations stop if max(abs(Cons_Values)/bas$Target)<Ctol, coded in ConsIter()}

\item{plt}{(boolean) for plotting}

\item{pltfreq}{number of iterations between plots}

\item{LAfreq}{storage frequency into list of coefficients LA}

\item{filter_strength}{weight of diagonal filter. i.e. 1+f/sqrt(M$bas$G.tk)}

\item{prn_ci}{(=FALSE) to print infos on constraint iterations}
}
\value{
membrane object with updates from MMC with data:

LA: list of recorded coefficients A

A: last coefficients

C: last curvature

E: last total energy

SDRC_Sample: recorded quantities (E,A,V,C, CN, IC, ID), where CN ist constraint norm, IC is constraint iteration count, and ID is the run-id.

SDRCiter; total SDRC steps, incl. previous calls

history: history of App-calls that created the result
}
\description{
Steepest Descend with Rosen Constraint Solver

Run SDRC on an existing membrane object to reduce energy by steepest descend.
Constraint violation is projected out by Rosens method (1961).
To stay near feasible points, constraint fix point iterations (Frickenhaus, Dissertation thesis, Humboldt-University Berlin, 1999) are performed every step.
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
data(M4,package = "MemRBC"); M<-M4
plot(M)
MemRBC_env$M.C0 <- 0
M <- SDRC(M, nsteps=10000, del_min=1e-6, LAfreq=100, cons_tol=1e-2, Gtol=1e-3)
plot(M)
attributes(last(M$LA)) 
\dontshow{\}) # examplesIf}
}
