% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{SEN}
\alias{SEN}
\title{SEN
compute stretch parameters (alpha, beta) from SEN}
\usage{
SEN(A, grd, bas, Ref, Wb_cur)
}
\arguments{
\item{A, grd, bas, Ref}{Coefficients, grid, basis and reference objects}

\item{Wb_cur}{: current E_SCM() result}
}
\value{
SEN object with alpha, beta and mean trace m of right Cauchy-Green
}
\description{
Compute the stretches alpha and shear beta from SEN parameters.
Needed before E_SEN is called.
}
