% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{SErenumbered_2_OBJ}
\alias{SErenumbered_2_OBJ}
\title{SErenumbered_2_OBJ
Read data from surface evolver dump file
 special case: faster version, but surface evolver has renumbered all before the dump}
\usage{
SErenumbered_2_OBJ(f.in, f.out, comment = "o")
}
\arguments{
\item{f.in, f.out}{: names of input SE dump /output .obj file}

\item{comment}{(="o") how Evolver input file encodes comments}
}
\value{
3d mesh object
}
\description{
SErenumbered_2_OBJ
Read data from surface evolver dump file
 special case: faster version, but surface evolver has renumbered all before the dump
}
