% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{SetConstraints}
\alias{SetConstraints}
\title{SetConstraints}
\usage{
SetConstraints(
  bas,
  Cons = c("gradA", "gradV"),
  QCons = c("Area", "Volume"),
  Target = c(140, 100),
  TNorm = c(140, 100)
)
}
\arguments{
\item{bas}{basis to modify constraints, e.g. M$bas (M MemRBC object)}

\item{Cons}{vector of character of constraint gradients, from "gradA","gradV" and "gradC"}

\item{QCons}{vector of constraint names, from "Area","Volume","Curv", same order as Cons}

\item{Target}{vector of constraint values, like c(140,100,88) for Area, Volume and Curv}

\item{TNorm}{vector of normalization constants for calculating the norm, i.e. degree of total constraint violation}
}
\description{
set the constraint target values and store information in basis.
Without parameters Cons, ..., the standard area and volume constraints are set with values (140, 100).
If you set a third constraint, for CNM you have to give M$Lambda a third component.
}
\examples{
data(D5); SetParams(D5)

# add curvature constraint:
SetConstraints(D5$bas,Cons=c("gradA","gradV","gradC"),
  QCons=c("Area","Volume","Curv"),
  Target=c(140,100,121),
  TNorm=c(140,100,121)) -> D5$bas  # store modified basis back into membranes D5 basis

# minimize with steepest descend under Rosen Constraint Projection

SDRC(D5,1000)->D5sdrc

# pair-plots of target quantities and energy E
plot(D5sdrc$SDRC_Sample[c("E","A","V","C")])

}
