% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{SetParam}
\alias{SetParam}
\title{set parameter in MemRBC envorinment}
\usage{
SetParam(x, value)
}
\arguments{
\item{x}{MemRBC parameter, e.g. "M.C0"}

\item{value}{value to store}
}
\value{
set value
}
\description{
set parameter in MemRBC envorinment
}
