% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{severe}
\alias{severe}
\title{severe for testing}
\usage{
severe(q1, q2, what = "some test", tol = 1e-12)
}
\arguments{
\item{q1, q2}{quantities to test for equality}

\item{what}{id for test_that call}

\item{tol}{(=1e-12) absolute tolerance for equality}
}
\description{
test two quantities for equality using `testthat::test_that()`
}
