% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{Sparsify}
\alias{Sparsify}
\title{Sparsify coefficient matrix}
\usage{
Sparsify(M)
}
\arguments{
\item{M}{membrane object to sparsify coefficients}
}
\value{
membrane object, with diagnosis results in $Sparse; drop_norm reports the L2-norm of all zeroed out coefficients.
}
\description{
zero out coefficients upto one (largest magnitude) entry per row (X,Y,Z).
For diagnosis, the energy values before and after Sparsify are printed and returned.
}
\examples{
data("M_stomatocyte_L12")
Sparsify(M_stomatocyte_L12) -> M_sparse
plot(M_sparse)
M_sparse$SparseE
}
