% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Spicules_as_Minima}
\alias{Spicules_as_Minima}
\title{locate Minima from quantity}
\usage{
Spicules_as_Minima(Q, grd, sgn = 1)
}
\arguments{
\item{Q}{spatial quantity, e.g. alpha from `SEN`, or energy density}

\item{grd}{grid as from a membrane objects `M$grd`}

\item{sgn}{(=+1) sign to test}
}
\value{
List of results, E for values, p for spatial positions
}
\description{
locate Minima from quantity
}
