% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{spreadout.uv}
\alias{spreadout.uv}
\title{spreadout.uv}
\usage{
spreadout.uv(uv)
}
\arguments{
\item{uv}{: n x 2 matrix of (u,v), like $uv returned from Brechbuhler}
}
\description{
Brechbuhler initial uv may be redistributed
Sparse regions are contracted.
}
