% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{StoreParams}
\alias{StoreParams}
\title{StoreParams}
\usage{
StoreParams(M)
}
\arguments{
\item{M}{Membrane to store actual parameters like M.C0, M.K_ADE etc. into}
}
\description{
Store current global membrane parameters in $Param the object; useful for load and save
}
\examples{
data("D5")
data("D6")
SetParams(D5) # set parameters stored in D5 to global variables
Storeparams(D6)->D6 # sets parameters from global variables to another object
# should be equivalent with
# D6$Params <- D5$Params
# but the SetParams(D5) sets global parameters for further use

}
