% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{synth}
\alias{synth}
\title{synth
DEPRECATE (same as synthX)
synthesize, but only for one spatial component in coefficients  A=A[,k] 
  for dim in 1..3: see synthX}
\usage{
synth(Y, A)
}
\arguments{
\item{Y, A}{basis function values matrix Y and a column of coefficient matrix A}
}
\value{
spatial coordinates
}
\description{
synth
DEPRECATE (same as synthX)
synthesize, but only for one spatial component in coefficients  A=A[,k] 
  for dim in 1..3: see synthX
}
