% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{synth_update}
\alias{synth_update}
\title{synth_update}
\usage{
synth_update(C, bas, i, k, del = 1e-06)
}
\arguments{
\item{C, bas}{coordinate and basis object}

\item{i}{spectral order as from (l,m)}

\item{k}{spatial dim.}

\item{del}{(=1e-6) : step length of update for forward difference}
}
\value{
changed entries in input coordinate object C
}
\description{
for forward differences, only add relevant del*Ylm
use this in Hessian computation by finite difference: 
    fwd: C=synth_update...(C,bas,i,k, +1e-6)
  reset: C=synth_update...(C,bas,i,k, -1e-6)
}
