% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{synthX}
\alias{synthX}
\title{synthesize X coordinates by matrix-matrix-product}
\usage{
synthX(Y, A)
}
\arguments{
\item{Y}{basis function value matrix, e.g. bas$Ylm}

\item{A}{coefficient matrix object}
}
\description{
synthesize X coordinates by matrix-matrix-product
}
