% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{thin_LA}
\alias{thin_LA}
\title{erase several attributes from coefficients
Only dim and dimnames are preserved.
If inversion matrix IM is attribute to M$A, it will be erased.}
\usage{
thin_LA(M)
}
\arguments{
\item{M}{MemRBC membrane object}
}
\value{
MemRBC object, with fewer attributes
}
\description{
erase several attributes from coefficients
Only dim and dimnames are preserved.
If inversion matrix IM is attribute to M$A, it will be erased.
}
\examples{
data(M4)
thin_LA(M4)->M4l
}
