% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{*.MemRBC}
\alias{*.MemRBC}
\title{scale coefficients by a scalar}
\usage{
\method{*}{MemRBC}(a, b)
}
\arguments{
\item{a, b}{scalar and MemRBC object for a*b$A (or the other way round)}
}
\value{
result MemRBC object with scaled coeffs
}
\description{
scale coefficients by a scalar
}
