% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{TotalEnergyDensity}
\alias{TotalEnergyDensity}
\title{TotalEnergyDensity}
\usage{
TotalEnergyDensity(S)
}
\arguments{
\item{S}{SEN, containing also h2cur$E_SCM_dens}
}
\value{
energy value
}
\description{
compute total energy density, to visualize
}
