% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{transplant}
\alias{transplant}
\title{transplant}
\usage{
transplant(M1, M2, plt = FALSE)
}
\arguments{
\item{M1}{membrane object to copy coefficients from}

\item{M2}{membrane object to copy coefficients into}

\item{plt}{(=FALSE) TRUE for a 3d plot of resulting shape}
}
\value{
modified M2 with new coefficients
}
\description{
copy coefficients from one membrane to another
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
data("M4",package = "MemRBC",envir=environment())
get_data_ZENODO("SS.rda")
data("SS",package = "MemRBC",envir=environment())
transplant(SS,M4)->SSmod
SSmod
\dontshow{\}) # examplesIf}
}
