% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{updateX}
\alias{updateX}
\title{updateX}
\usage{
updateX(A, grd, bas)
}
\arguments{
\item{A}{coefficients of shape}

\item{grd}{grid from on which the basis is computed}

\item{bas}{basis function values $Ylm and their derivatives, like $Ylm_u.}
}
\value{
basis object, type is list.
}
\description{
updates coordinates C$X and their partial derivatives wrt. u,v, like C$X_u.
}
