% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{vectoarr_cxx}
\alias{vectoarr_cxx}
\title{needed in c++}
\usage{
vectoarr_cxx(x, ndof, Aimax)
}
\arguments{
\item{x, ndof, Aimax}{build a ndof x Aimax x 3 array from values x}
}
\value{
resulting array
}
\description{
needed in c++
}
