% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{vectomat_cxx}
\alias{vectomat_cxx}
\title{needed in c++}
\usage{
vectomat_cxx(x, Aimax)
}
\arguments{
\item{x, Aimax}{build a Aimax x 3 matrix from values x}
}
\value{
resulting matrix
}
\description{
needed in c++
}
