% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{VertexAreasOBJ}
\alias{VertexAreasOBJ}
\title{build vertex areas from triangle areas using Rvcg
ATTENTION: if O contains double points, the resulting area vector needs
masking, i.r. `av<-av[-bas$mask]` 
Furthermore it may be neccessary to close the object O at the v=0|2pi parameter line 
favour of correct area evaluations via Rvcg}
\usage{
VertexAreasOBJ(O)
}
\arguments{
\item{O}{mesh3d object}
}
\value{
vector of areas per vertex
}
\description{
build vertex areas from triangle areas using Rvcg
ATTENTION: if O contains double points, the resulting area vector needs
masking, i.r. `av<-av[-bas$mask]` 
Furthermore it may be neccessary to close the object O at the v=0|2pi parameter line 
favour of correct area evaluations via Rvcg
}
