% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Weighted_FitAlm}
\alias{Weighted_FitAlm}
\title{Weighted SPHARM fit as from https://www.researchgate.net/publication/228570513_Weighted_spherical_harmonic_representation_and_its_application_to_cortical_analysis
Chung, Moo & Shen, Li & Dalton, Kim & Kelley, Daniel & Robbins, Steven & Evans, Alan & Davidson, Richard. (2006). Weighted spherical harmonic representation and its application to cortical analysis.}
\usage{
Weighted_FitAlm(X, bas, sigma = 0.001, CL = 0.95)
}
\arguments{
\item{X, bas}{spatial coords and basis}

\item{sigma}{(=0.001) sigma paraleter of weight decrease with spectral l}

\item{CL}{(=0.95) confidence level to test for significant terms}
}
\value{
fitted coefficients
}
\description{
Weighted SPHARM fit as from https://www.researchgate.net/publication/228570513_Weighted_spherical_harmonic_representation_and_its_application_to_cortical_analysis
Chung, Moo & Shen, Li & Dalton, Kim & Kelley, Daniel & Robbins, Steven & Evans, Alan & Davidson, Richard. (2006). Weighted spherical harmonic representation and its application to cortical analysis.
}
