% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Weighted_FitAlm_L3}
\alias{Weighted_FitAlm_L3}
\title{Weighted_FitAlm_L3(X,bas,sigma)}
\usage{
Weighted_FitAlm_L3(X, bas, sigma = 0.001, CL = 0.95)
}
\arguments{
\item{X}{spatial coordinates at spherical coords U,V used for bas}

\item{bas}{basis functions precomputed from U,V corresponding to vertices}
}
\value{
list with coeffs "A", smoothed "Xestim",
}
\description{
kernel regularized least squares fit of Fourier coefficients Alm to vertex coordinates
first fitted block is l=0...3; any l=0 coeffient (=translation dof) is dropped on return
}
\examples{
# X4 is a surface object (of genus 0) from rgl::writeOBJ()
data(X4)
BB <- Brechbuehler.Init.uv.2(X4) # not well documented
spreadout.uv(BB$uv)->uv
bas <- MakeBasis_UV(L_max=12, uv[,1], uv[,2])
A <- Weighted_FitAlm_L3(Obj2X_centre(O), bas, 0.001)
G <- MakeGrid_GaussLegendreSimpson(n=30)
M <- structure(class="MemRBC",list(bas=bas,grd=G,A=A))
plot(M) # problem to fit an armed structure with Brechbühlers initial (u,v)
...
}
