% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{X2Obj}
\alias{X2Obj}
\title{make coordinates from object}
\usage{
X2Obj(O, X)
}
\arguments{
\item{O}{mesh3d object to insert data}

\item{X}{coordinates to store}
}
\value{
mesh3d object
}
\description{
make coordinates from object
}
