% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{X2ObjQ}
\alias{X2ObjQ}
\title{X2ObjQ}
\usage{
X2ObjQ(O, X)
}
\arguments{
\item{O}{3d mesh object}

\item{X}{3d-coordinates, like from a coord object in `C$X`}
}
\value{
3d object with quads
}
\description{
puts coordinates X into object O, which may be of quads
}
