% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{X_center_masked}
\alias{X_center_masked}
\title{X_center_masked(X,bas)}
\usage{
X_center_masked(X, bas)
}
\arguments{
\item{X}{spatial coordinates at spherical coords U,V used for bas}

\item{bas}{basis functions precomputed from U,V corresponding to vertices}
}
\value{
centered X
}
\description{
remove center of mass by translation
uses masked X to avoid double counting v=0 and v=2pi points
}
\examples{
N=50
g=MakeGrid_GaussLegendreSimpson(N)
b=MakeBasis_UV(8,g$U,g$V)
a=MakeSphere(g,b)
X=updateX_only(a,g,b)$X # initial sphere data
X<-invag_N(X,g,0.3,1)
rgl::plot3d(X_center_masked(X))
}
