% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALM.R
\name{ALM}
\alias{ALM}
\title{Augmented Lagrangian Minimization}
\usage{
ALM(
  M,
  curv,
  nsteps = 10,
  plt = TRUE,
  method = "SLSQP",
  maxiter_solver = 100,
  ncores = 4,
  LAfreq = 5
)
}
\arguments{
\item{M}{The input membrane with initial data and reference}

\item{curv}{target curvature}

\item{nsteps}{(=10) number of MMC steps to be run}

\item{plt}{(boolean) for plotting}

\item{method}{"SLSQP" for internal nloptr minimizer}

\item{maxiter_solver}{(=100) max. solver iterations}

\item{ncores}{(=4) number of parallel cores tu use}

\item{LAfreq}{(=5) frequency of recording coeffs A to LA}
}
\value{
membrane object with updates from MMC with data:

LA: list of recorded coefficients A

A: last coefficients

M.mu: last mu value

M.lam: last lambda vector

history: history of App-calls that created the result
}
\description{
run ALM on an existing membrane object to adjust shape to target curvature, minimizing the residual norm RN.
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
MemRBC_env$M.C0<-0
M <- MakeStandardRBC(L=3,C0=0)
plot(M,alpha=0.6)
 M <- ALM(M,curv=Curv(M)+2,10,ncores=2)
plot(M,alpha=0.6,col="red")
M
\dontshow{\}) # examplesIf}
}
