% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Brechbuehler.Init.uv.2}
\alias{Brechbuehler.Init.uv.2}
\title{Brechbuehler.Init.uv.2}
\usage{
Brechbuehler.Init.uv.2(
  X1,
  Fit_order = 12,
  InitFit = FALSE,
  poles.axis = 2,
  mat.mode = c("cotan.lapl", "cotan.chi", "euclid"),
  file.out = "Brechbuehler-init-uv-2.obj",
  rxy = 0.01
)
}
\description{
computes spherical coordinates (u,v) for vertices of a 3D-object X
The original algorithm from Brechbühler (1995) is implemented,
but spherical areas are not iterated for refinement.
The resulting (u,v) may not be optimal for fitting.
}
