% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNM.R
\name{CNM}
\alias{CNM}
\title{Constraint Newton Minimizer}
\usage{
CNM(
  M,
  nsteps = 5,
  del = 0.3,
  del_control = FALSE,
  diag.reg = 0,
  pinv = TRUE,
  LAfreq = 1,
  Mtol_Newton = 1e-04,
  ncores = 4,
  keepEig2A = FALSE,
  plt = TRUE,
  cluster = FALSE
)
}
\arguments{
\item{M}{The input membrane with initial data and reference}

\item{nsteps}{(=5) number of Newton steps to be performed}

\item{del}{(=0.3) update factor delte (<1)}

\item{del_control}{(=FALSE) TRUE for use of statoíc step length del}

\item{diag.reg}{(=0) diagonal regularization of to be inverted matrix}

\item{pinv}{(=TRUE) use pseudo-inverse in solver}

\item{LAfreq}{(=1) storage frequency into list of coefficients LA}

\item{Mtol_Newton}{(=1e-4) stopping criterion for KKT norm and constraints norm}

\item{ncores}{(=4) for parallel Hessian on ncores CPU-cores}

\item{keepEig2A}{(=FALSE) to keep eigensystem as attribute "EigsH" in $A (and thus LA[[]] -> high memory demands)}

\item{plt}{(=TRUE) for plotting two figures with alpha/beta stress-shear parameters}

\item{cluster}{(=FALSE) for parallel Hessian using a cluster with ncores processes}
}
\value{
membrane object with updated data from CNM:

LA : list of recorded coefficients A

A : last coefficients

history : history of App-calls that created the result

E_CNM : vector of nsteps energy values

CNMiter : number of iterations, incl. previous calls

proc_time : aggregated processing time of all Apps called before with this object

Eig.H : eigensystem of last energy+constraints Hessian

Grad : final energy gradient

Jacobian : constraint Jacobian

Hessian : full energy Hessian matrix

ConsHessians : Lagrangian Hessians from active constraints
}
\description{
run CNM on an existing membrane object to minimize energy under exact constraints on area and volume
You may include curvature constraint as well.
Temperature is controlled by parameter kT.
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
data(M4)
MemRBC_env$M.C0 <- -4
plot(M4)
M4<-SDRC(M4,10,del_min=3e-6)
M4 <- CNM(M4,nsteps=20,cluster=TRUE,ncores=4,del_control=TRUE)
plot(M4)
M
\dontshow{\}) # examplesIf}
}
