% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNM.R
\name{CNM}
\alias{CNM}
\title{Constraint Newton Minimizer}
\usage{
CNM(
  M,
  nsteps = 5,
  del = 0.3,
  diag.reg = 0,
  LAfreq = 1,
  Mtol_Newton = 1e-04,
  thresh = 1e-04,
  ncores = 4,
  plt = TRUE,
  filter_delta = ID,
  cluster = FALSE
)
}
\arguments{
\item{M}{The input membrane with initial data and reference}

\item{nsteps}{(=5) number of Newton steps to be performed}

\item{del}{(=0.3) update factor delte (<1)}

\item{LAfreq}{(=1) storage frequency into list of coefficients LA}

\item{Mtol_Newton}{(=1e-4) stopping criterion for norm of gradient}

\item{thresh}{(=1e-4)threshold for pseudoinverse to suppress almost zero eigenvalues in matrix inversion}

\item{ncores}{(=4) for parallel Hessian with ncores}

\item{plt}{(=TRUE) for plotting two figures with alpha/beta stress-shear parameters}

\item{cluster}{(=FALSE) for parallel Hessian using a cluster with ncores processes}
}
\value{
membrane object with updated data from CNM:

LA: list of recorded coefficients A

A: last coefficients

history: history of App-calls that created the result

E_CNM: vector of nsteps energy values

CNMiter: number of iterations, incl. previous calls

proc_time: aggregated processing time of all Apps called before with this object
}
\description{
run CNM on an existing membrane object to minimize energy under exact constraints on area and volume
You may include curvature constraint as well.
Temperature is controlled by parameter kT.
}
\examples{
M <- MakeStandardRBC(L=5)
plot(M)
M <- MMC(M,nsteps=10000,kT=0.00411,kTfac=0.99,kTfreq=100)
plot(M)
M
}
