% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{ConsIter}
\alias{ConsIter}
\title{ConsIter}
\usage{
ConsIter(
  A,
  grd,
  bas,
  C,
  g2,
  Ctol = 0.001,
  nsteps = 20,
  prn = FALSE,
  del_cons = 0.3,
  do_one = TRUE
)
}
\arguments{
\item{A, grd, bas, C}{MemRBC parameters coeff, grid, basis and coords}

\item{g2}{result of `E_SCM()`}

\item{Ctol}{(=1e-3) constraint tolerance}

\item{nsteps}{(=20) maximum interation count}

\item{prn}{(=FALSE) no printing of iterations}

\item{del_cons}{(=0.3) delta parameter of fixed-point iteration}

\item{do_one}{(=TRUE) at least perform first iteration}
}
\value{
list with iterated coefficients A, h2 for E_SCM, g2 for Grad_SCM
}
\description{
Iterates coefficients to fulfill constraints.
Needed in Rosen Projection methods.
only SCM gradient is computed for constraint Jacobian
}
