% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{ConsRHS}
\alias{ConsRHS}
\title{ConsRHS}
\usage{
ConsRHS(h2, bas)
}
\arguments{
\item{h2}{current energy from `E_SCM`}

\item{bas}{basis}
}
\value{
a named vector of constrained violations, i.e. g_i(x)
}
\description{
Returns constraint violation values, e.g. for rhs of Rosen solver
}
