% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{ConstraintHessian}
\alias{ConstraintHessian}
\title{ConstraintHessian}
\usage{
ConstraintHessian(H, bas, Lambda)
}
\arguments{
\item{H}{hessian matrix}

\item{bas}{basis function data like M$bas}

\item{Lambda}{Lgrangian multipliers, to check number of constraints}
}
\value{
bordered Hessian (Hessian with Jacobian blocks)
}
\description{
constraint (=bordered) Hessian for `CNM()` Newton minimizer
}
