% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{E_FullModel_Penalty}
\alias{E_FullModel_Penalty}
\title{unified model energy with penalties
 @description
 $bas$Nc and $bas$Cons are used to address active constraints}
\usage{
E_FullModel_Penalty(A, grd, bas, Ref)
}
\arguments{
\item{A, grd, bas, Ref}{standard parameters from MemRBC object}
}
\value{
full penalized energy
}
\description{
unified model energy with penalties
 @description
 $bas$Nc and $bas$Cons are used to address active constraints
}
