% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{FitAlm}
\alias{FitAlm}
\title{FitAlm}
\usage{
FitAlm(X, bas, WX = rep(1, nrow(X)))
}
\arguments{
\item{X, bas}{input data and basis; bas$mask must be set to ecluded X points indices}

\item{WX}{(=1) spatial weights, could be sin(grd$U)}
}
\description{
fit coefficients from 3d-coordinates
}
