% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{FitDiscocyte_L5}
\alias{FitDiscocyte_L5}
\title{FitDiscocyte_L5}
\usage{
FitDiscocyte_L5(C0 = 0, V0 = 100)
}
\arguments{
\item{C0}{C0-value to use for initial minimization}
}
\value{
membrane object
}
\description{
Fit a discocyte from data, spectral order L=5
}
\examples{
FitDiscocyte_L5(C0=0.5)->M
plot(M)
}
