% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{FitFast}
\alias{FitFast}
\title{FitFast}
\usage{
FitFast(bas, X)
}
\arguments{
\item{bas, X}{basis with `bas$IM`, and X with 3d-coordinates to fit}
}
\value{
fitted coefficients
}
\description{
fits coefficients to X from least squares solver matrix IM in bas
This is used for fitting e.g. displacements
like in pertA_complex.
ATTENTION: bas$mask is needed for masking double coords.
If not there or not meaningful, set to a default excludion of first spatial point by bas$mask=1
}
