% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{FitProlate_Ellipsoid_L5}
\alias{FitProlate_Ellipsoid_L5}
\title{Fit coefficients to a prolate ellipsoid}
\usage{
FitProlate_Ellipsoid_L5(C0 = 2.562, V0 = 85.66)
}
\arguments{
\item{C0}{(=2.562) Undustick spontaneous curvature}

\item{Vp}{(=85.66) Undustick volume, which is reduced volume 0.55}
}
\value{
mebrane object of prolate shape to be optimized further, e.g. for Undustick
}
\description{
Fit an ellipsoid of prolate shape e.g. with Undustick parameters.
Global parameters are set for Lipid modeling, i.e. SEN is switched off.
}
\examples{
FitProlate_Ellipsoid_L5() -> M
plot(M)

}
