% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{FitStomatocyte_L}
\alias{FitStomatocyte_L}
\title{FitStomatocyte_L}
\usage{
FitStomatocyte_L(L = 7, C0 = -3, V0 = 100, dt = 0.001)
}
\arguments{
\item{L}{spectral order of output membrane}

\item{C0}{(=-3) C0-value to use for initial minimization}

\item{V0}{(=100) target value of volume}

\item{dt}{(=1e-3) time step in PNEM minimizer}
}
\value{
membrane object
}
\description{
Fit a stomatocyte shape from data, spectral order L
}
\examples{
FitStomatocyte_L(L=4,C0=0.5)->M
plot(M)
}
