% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{FullModelHessian}
\alias{FullModelHessian}
\title{FullModelHessian}
\usage{
FullModelHessian(A, grd, bas, Ref, del = 1e-06, Ctol = 0.001)
}
\arguments{
\item{del}{finite difference delta for coefficients}
}
\value{
Full Hessian matrix, ie also constraint Jacobian
}
\description{
computes the full model Hessian
of the coefficients, as used in CNM.
It is based on finite differences of gradients
with a symmetrization.
}
