% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{FullModelHessian_Par}
\alias{FullModelHessian_Par}
\title{FullModelHessian_Par}
\usage{
FullModelHessian_Par(
  A,
  grd,
  bas,
  Ref,
  del = 1e-06,
  Mem_mc.cores = 4,
  timing = TRUE,
  stopdown = TRUE,
  cl = NULL
)
}
\arguments{
\item{A, grd, bas, Ref}{: usual membrane data coeffs, grid, basis, SEN-reference}

\item{del}{: forward difference length parameter}

\item{Mem_mc.cores}{(=4) for parallel cluster size}

\item{timing}{(=TRUE) output some timing information}

\item{stopdown}{(=TRUE) if cluster should be stopped before exit.}

\item{cl}{(=NULL) : parallel cluster-ID from previous result (`H$cl` in `CNM()`)}
}
\value{
Full Hessian matrix, ie also constraints, but not bordered
}
\description{
parallel computed Hessian including constraint gradients; serial gradients but a list-parallel approach to assemble H
}
