% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{GEMINI_cut_mesh_along_path}
\alias{GEMINI_cut_mesh_along_path}
\title{Cut a Mesh Along a Path of Vertices}
\usage{
GEMINI_cut_mesh_along_path(mesh, path)
}
\arguments{
\item{mesh}{An Rvcg mesh3d object.}

\item{path}{A vector of integer vertex indices representing the cut path.
Must be a connected sequence of edges.}
}
\value{
A new mesh3d object with the cut applied (more vertices, updated faces).
}
\description{
Opens a closed mesh by "unzipping" it along a sequence of vertices.
Interior vertices of the path are duplicated. Endpoints remain as "hinges".
}
