% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{GEMINI_disk_conformal_map}
\alias{GEMINI_disk_conformal_map}
\title{GEMINI_disk_conformal_map}
\usage{
GEMINI_disk_conformal_map(mesh, L = NULL, plt = FALSE, spherical = TRUE)
}
\arguments{
\item{mesh}{An Rvcg mesh3d object (must be an open mesh, not a closed sphere)}

\item{L}{The mesh Laplacian matrix. If NULL, it is computed.}

\item{plt}{(=FALSE) to plot}

\item{spherical}{(=TRUE) for giving not u,v on a disk but spherical coordinates (u,v)}
}
\value{
List containing the 2D coordinates uv and the flattened 3d mesh mesh_flat.
}
\description{
Map a 3D Mesh to a 2D Unit Circle (Disk Conformal Map)
}
