% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{GEMINI_Intersect_Mem_Mem}
\alias{GEMINI_Intersect_Mem_Mem}
\title{GEMINI_Intersect_Mem_Mem}
\usage{
GEMINI_Intersect_Mem_Mem(M1, M2 = M1)
}
\arguments{
\item{M1, M2}{(M2=M1 for self-intersection) MemRBC membrane objects}
}
\value{
TRUE, if M1 and M2 are intersecting; else: FALSE
}
\description{
compute from membrane data the 3D objects and check for intersection
(experimental, not checked intensively).
}
