% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{GEMINI_Obj_Obj_Intersect}
\alias{GEMINI_Obj_Obj_Intersect}
\title{GEMINI_Obj_Obj_Intersect}
\usage{
GEMINI_Obj_Obj_Intersect(S1, S2 = S1)
}
\arguments{
\item{S1, }{S2 MemRBC objects; S2=S1 is used for self-intersections}
}
\value{
boolean, TRUE if intersection is probable
}
\description{
this code from GEMINI intends to locate spatial intersections of 3D-objects
}
