% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Grad_FullModel_Penalty}
\alias{Grad_FullModel_Penalty}
\title{unified gradient computation with penalties
 @description
 $bas$Nc and $bas$Cons are used to address active constraints}
\usage{
Grad_FullModel_Penalty(A, grd, bas, Ref, S)
}
\arguments{
\item{A, grd, bas, Ref}{standard parameters from MemRBC object}

\item{S}{from SEN()}
}
\value{
Gradient of full penalized energy in coefficient space, vector
}
\description{
unified gradient computation with penalties
 @description
 $bas$Nc and $bas$Cons are used to address active constraints
}
