% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{imag.delta.aligned}
\alias{imag.delta.aligned}
\title{plot svd-aligned objects}
\usage{
imag.delta.aligned(A1, A2, grd2, bas2, shade2 = FALSE)
}
\arguments{
\item{A1, A2}{input coefficients}

\item{grd2, bas2}{grid and basis of A2}

\item{shade2}{(=FALSE) to draw shade3d of second shape}
}
\value{
List with obj1, and obj2 aligned 3d-objects and coordinate objects C1, C2
}
\description{
svd alignment of shape coefficients into 3d-objects, see
https://www.cse.wustl.edu/~taoju/cse554/lectures/lect07_Alignment.pdf
and plotting of spatial difference vectors
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
data("M4",package = "MemRBC")
M4p=M4
M4p$A=M4$A+rnorm(M4$bas$Ai_max*3,sd=0.4)
imag.delta.aligned(M4p$A,M4$A,M4$grd,M4$bas)->L
rgl::open3d()
rgl::plot3d(L$obj1,alpha=0.6,col=1,aspect=FALSE)
rgl::plot3d(L$obj2,alpha=0.6,col=2,add=TRUE)
\dontshow{\}) # examplesIf}
}
