% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{imag.delta.arrowws.pca}
\alias{imag.delta.arrowws.pca}
\title{plot arrows of change between two shape states}
\usage{
imag.delta.arrowws.pca(A1, A2, grd2, bas2, O1)
}
\arguments{
\item{A1, A2}{coefficients for the shapes}

\item{grd2, bas2}{grid and basis of second shape A2}

\item{O1}{3d object for A1}
}
\description{
plot arrows of change between two shape states
}
