% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{imag}
\alias{imag}
\title{imag}
\usage{
imag(field, grd, nx = grd$nv, ny = grd$nu/2, ...)
}
\arguments{
\item{field}{scalar field, size of grd$nv x grd$nu}

\item{grd}{the grid of (u,v) on which the scalar is defined}
}
\description{
plot a scalar in 2D
}
