% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{image.MemRBC}
\alias{image.MemRBC}
\title{image}
\usage{
\method{image}{MemRBC}(M, q = M$SEN$alpha, main = expression(alpha), ...)
}
\arguments{
\item{M}{Membrane to use for shape plot}

\item{q}{a 2D field to project onto shape, eg. M$SEN$alpha}
}
\description{
2D plot of color coded variable on the shape
}
