% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{MakeBasis_UV}
\alias{MakeBasis_UV}
\title{MakeBasis_UV}
\usage{
MakeBasis_UV(
  L_max = 4,
  u,
  v,
  Pointsymmetry = FALSE,
  only_Ylm = FALSE,
  kind = "Ylm",
  KleinBottle = FALSE
)
}
\description{
compute the basis functions on given (u,v) values
(u,v) are usually from grd$U, grd$V, but you may also use irregular (u,v)
for example for fitting coefficients to a 3d-object, in which (u,v)-values
may be given as texture coordinates texccords.
}
\details{
@param L_max (=4) spectral order of basis
 @param u,v vectors of spehrical angles u (in 0..pi) and v (in 0..2*pi).
 @param only_Ylm (=FALSE) for excluding derivative computation.
 @return basis, to be stored as $bas in membrane object or for other use.
 The basis also contains standard constraints for area and volume, which may be modified with SetConstraints()
}
