% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeStandardRBC.R
\name{MakeInvaginated}
\alias{MakeInvaginated}
\title{Create a invaginated cell from a sphere
via invag_N() positive depth.
NOTE: spatial points are weighted by sin(u) for the fit}
\usage{
MakeInvaginated(
  w = 0.38,
  d = 1.4,
  r = 1.045,
  f = 0.75,
  uv = c(0, 0),
  plt = FALSE
)
}
\arguments{
\item{w}{(=0.38) width of grove}

\item{d}{(=1.4) depth of grove}

\item{r}{(=1.045) radius to tune resulting volume}

\item{f}{(=0.75) factor for coeff A[2,3]}

\item{uv}{(=c(0,0)) rotation before invagination}

\item{plt}{(=FALSE) for control plot}
}
\value{
MemRBC object with invagination coded in the coefficients $A
}
\description{
Create a invaginated cell from a sphere
via invag_N() positive depth.
NOTE: spatial points are weighted by sin(u) for the fit
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
MakeInvaginated() -> ivs
ivs
ivs$bas$Target
rgl::open3d()
plot(ivs)
MemRBC_env$M.C0 <- -6
MMC(ivs,100000,plt=TRUE,pltfreq=100,LAfreq=1000,C0=-6) -> ivs_mmc
\dontshow{\}) # examplesIf}
}
