% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{MakeMemRBC}
\alias{MakeMemRBC}
\title{MakeMemRBC}
\usage{
MakeMemRBC(A, G, B)
}
\arguments{
\item{A, G, B}{coeffs, grid and basis objects for MemRBC object}
}
\value{
MemRBC object
}
\description{
Build a MemRBC object from coefficients, grid and basis
}
