% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{MemPCA}
\alias{MemPCA}
\title{MemPCA}
\usage{
MemPCA(M, WX = rep(1, M$grd$ndof))
}
\arguments{
\item{M}{the input membrane to be rotated}
}
\value{
membrane object with new coefficients of PCA-rotated spatial coordinates.
}
\description{
rotate Membrane coefficients to coordinates of principle axes of inertia.
Remark: the reference SEN in $Ref remains unchanged.
}
\examples{
data("M_stomatocyte_L12"); SetParams(M_stomatocyte_L12)
plot(M_stomatocyte_L12)
MemPCA(M_stomatocyte_L12)->M2
plot(M2)
}
