% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{MemPCA}
\alias{MemPCA}
\title{MemPCA}
\usage{
MemPCA(M, WX = rep(1, M$grd$ndof))
}
\arguments{
\item{M}{the input membrane to be rotated}

\item{WX}{(=1) spatial weights for the fit of coefficients, try sin(grd$U) to downweight pole influence}
}
\value{
membrane object with new coefficients of PCA-rotated spatial coordinates.
}
\description{
rotate Membrane coefficients to coordinates of principle axes of inertia.
Remark: the reference SEN in $Ref remains unchanged.
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
data("M4",envir=environment())
MemPCA(M4)->M2
plot(M2)
\dontshow{\}) # examplesIf}
}
