% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{MemStab}
\alias{MemStab}
\title{MemStab}
\usage{
MemStab(
  M,
  mc.cores = 4,
  plt = FALSE,
  plt_mode = 1,
  plt_scale = 0.5,
  serial = FALSE
)
}
\value{
M$Stab with eigensystem of Hessian of energy without constraints
}
\description{
analyse membrane shape stability in terms of eigenvectors of Hessian
}
