% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMC.R
\name{MMC}
\alias{MMC}
\title{run MMC on an existing membrane object under penalty constraints for volume and area.
Every 75 rejected steps, the interval size of random numbers added on coefficients is decreased by 5%.
If the interval size is below 1e-6, it is re-initialized to 0.01.}
\usage{
MMC(
  M,
  nsteps = 1000,
  plt = FALSE,
  pltfreq = 5,
  LAfreq = 200,
  sd = 0.004,
  kT = 0.00411,
  nm = FALSE,
  kTfac = 1,
  kTfreq = 100,
  filter = ID
)
}
\arguments{
\item{M}{The input membrane with initial data and reference}

\item{nsteps}{number of MMC steps to be run}

\item{plt}{(=FALSE) for plotting}

\item{LAfreq}{(=200) for storing shape coefficients every LAfreq accepted steps}

\item{sd}{(=0.004) standard deviation parameter to control steps}

\item{kT}{(=0.00411) Boltzmann-constant * temperature energy scale}

\item{nm}{(=FALSE) normal motion filter}

\item{kTfac}{(=1.0) cooling factor (0<kTfac<1) or heating factor (kTfac>1)}

\item{kTfreq}{(=100) number of accepted steps between two scalings of kT by kTfac}

\item{filter}{(=ID) use ApplyFlt_L1 for fixed overall orientation in L1-coeffs; requires SetFlt_L1(M) before}

\item{plfreq}{(=5) for plotting shape every plotfreq number of accepted steps}
}
\value{
membrane object with updates from MMC with data:

LA: list of recorded coefficients A

A: final coefficients

kT:  value of kT on exit, important for restarting

MMCacceptanceRate: acceptance rate of the call

MMCiter: total MMC steps, incl. MMC from previous calls

history: history of App-calls that created the result
}
\description{
run MMC on an existing membrane object under penalty constraints for volume and area.
Every 75 rejected steps, the interval size of random numbers added on coefficients is decreased by 5%.
If the interval size is below 1e-6, it is re-initialized to 0.01.
}
\examples{
M <- MakeStandardRBC(L=5)
plot(M)
#  annealing simulation (decrease kT by kTfac every Ktfreq accepted steps)
M <- MMC(M,nsteps=100000, kT=0.00411, kTfac=0.99, kTfreq=100)
plot(M)
M
}
