% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMCC.R
\name{MMCC}
\alias{MMCC}
\title{MMCC}
\usage{
MMCC(
  M,
  curv = Curv(M),
  nsteps = 1000,
  plt = FALSE,
  pltfreq = 5,
  LAfreq = 200,
  sd = 0.004,
  kT = 0.00411,
  pertA = pertA_Unif,
  kTfac = 1,
  kTfreq = 100
)
}
\arguments{
\item{M}{The input membrane with initial data and reference}

\item{curv}{the constraint curvature (default is current curvature of M)}

\item{nsteps}{number of MMC steps to be run}

\item{plt}{(boolean) for plotting}

\item{LAfreq}{for storing shape coefficients every LAfreq accepted steps}

\item{sd}{standard deviation parameter to control steps}

\item{kT}{Boltzmann-constant * temperature energy scale}

\item{kTfac}{cooling factor}

\item{kTfreq}{frequency for cooling factor applied on kT in units of accepted steps}

\item{plofreq}{for plotting shape every plotfreq number of accepted steps}

\item{nm}{(boolen) normal motion filter}
}
\value{
membrane object with updates from MMC with data:

LA: list of recorded coefficients A

A: final coefficients

kT:  value of kT on exit, important for restarting

MMCCacceptanceRate: acceptance rate of the call

MMCCiter: total MMCC steps, incl. MMCC from previous calls

history: history of App-calls that created the result
}
\description{
Metropolis Montecarlo with Curvature Constraint

run MMC on an existing membrane object under penalty constraints for volume, area and curvature
}
\examples{
M <- MakeStandardRBC(L=5)
plot(M)
#  annealing simulation (decrease kT by kTfac every Ktfreq accepted steps)
M <- MMCC(M, curv=Curv(M)+0.25, nsteps=100000, kT=0.00411, kTfac=0.99, kTfreq=100)
plot(M)
M
}
