% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{Movie}
\alias{Movie}
\title{Movie}
\usage{
Movie(
  M,
  x = 400,
  y = 400,
  from = 1,
  to = length(M$LA),
  sleep = 0.025,
  phi = pi/100,
  phi1 = phi/2,
  ...
)
}
\description{
show shapes along recorded coefficients in M$LA
}
