% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{NorthSouth}
\alias{NorthSouth}
\title{find two extremal vertices along PC1 and some 5 paths to select from}
\usage{
NorthSouth(O, w = 1)
}
\arguments{
\item{O}{3d object}
}
\value{
NS=c(N,S), North and South vertex indices

Path: list of index vectors connecting N and S by k-shortest path
}
\description{
find two extremal vertices along PC1 and some 5 paths to select from
}
\examples{
data(SF4lr)
SF4lr->O
NorthSouth(O,13)->P # take 13th path
P$Paths # candidate paths of vertices for cutting
P$LVPath # 13th lowest var(z) path
PlotPaths(O,P)
O1 <- GEMINI_cut_mesh_along_path(O,P$LVPath)
}
