% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMC.R
\name{pertA_complex}
\alias{pertA_complex}
\title{pertA_complex}
\usage{
pertA_complex(A, bas, sd, n, nn = 12)
}
\description{
Perturb with gaussian along spatial normals.
 BEFORE USE:
a) Normals of neighbours require to fill global M.nn=Nuv(bas$uv)
b) the basis mus contain a matrix IM for least squares inversion in FitFast();
Use bas<-MakeIM(bas,WX=sin(grd$U)) to create IM.
}
