% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMC.R
\name{pertA_rndX}
\alias{pertA_rndX}
\title{pertA_rndX}
\usage{
pertA_rndX(A, bas, sd, n, sample_fraction = 0.65)
}
\arguments{
\item{A, bas}{coeffs and basis}

\item{sd}{standard dev}

\item{n}{not used}

\item{sample_fraction}{(=0.65) fraction of points to be displaced}
}
\value{
changed coeffs A+dA
}
\description{
Perturb coeffs from spatial coordinates.
 BEFORE USE:
the basis mus contain a matrix IM for least squares inversion in FitFast();
Use bas<-MakeIM(bas,WX=sin(grd$U)) to create IM.
}
