% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{plot.MemRBC}
\alias{plot.MemRBC}
\title{plot}
\usage{
\method{plot}{MemRBC}(x, wire = TRUE, wire_col = "black", ...)
}
\arguments{
\item{x}{Membrane to use for shape plot}

\item{wire}{(boolean) FALSE if  to suppress mesh quads plotted}

\item{wire_col}{(="black")}

\item{...}{further plotting options, e.g. col="green"}
}
\description{
3D plot of membrane shape
}
