% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{plot.MemRBC}
\alias{plot.MemRBC}
\title{plot}
\usage{
\method{plot}{MemRBC}(R, wire = TRUE, wire_col = "black", ...)
}
\arguments{
\item{wire}{(boolean) FALSE if  to suppress mesh quads plotted}

\item{...}{further plotting options, e.g. col="green"}

\item{M}{Membrane to use for shape plot}
}
\description{
3D plot of membrane shape
}
