% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{plot3a}
\alias{plot3a}
\title{plot3a}
\usage{
plot3a(
  X,
  grd,
  pnts = FALSE,
  clip = FALSE,
  col = "black",
  alpha = 1,
  cont = TRUE,
  cont.grid = FALSE,
  fill = TRUE,
  fn = "z",
  fn_data = "z",
  ...
)
}
\arguments{
\item{X}{Coordiates, e.g. in C$X from updateX()}

\item{grd}{grid with a basic rgl-object grd$Obj}
}
\description{
plot coordinates C$X (after updateX)
}
\examples{
data("M1")
 # take required data from M1
plot3a(updateX(M1$A,M1$grd,M1$bas,M1$)$X,M1$grd)

}
