% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{plot3q}
\alias{plot3q}
\title{plot3q}
\usage{
plot3q(X, grd, col = "black", alpha = 1, ...)
}
\arguments{
\item{X}{Coordiates, e.g. in C$X from updateX()}

\item{grd}{grid with a basic rgl-object grd$Obj}
}
\description{
plot coordinates C$X (after updateX) with wireframe of quadrilaterals
}
\examples{
data("M1")
 # take required data from M1
plot3q(updateX(M1$A,M1$grd,M1$bas,M1$)$X,M1$grd)
}
