% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{plot3qs}
\alias{plot3qs}
\title{plot3qs}
\usage{
plot3qs(X, grd, s, specular = "black", pal = heat.colors, ...)
}
\arguments{
\item{X}{Coordiates, e.g. in C$X from updateX()}

\item{grd}{grid with a basic rgl-object grd$Obj}

\item{s}{scalar to plot as color code on shape}

\item{specular}{(="black") material reflectance}

\item{pal}{(=heat.colors) color palette to use}

\item{...}{further params to shade3d(), e.g. alpha=0.5 for transparency}
}
\value{
-
}
\description{
plot with coordinates C$X (after updateX) and scalar as color code
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
data("M4",package = "MemRBC"); 
SetParams(M4)
 update(M4,"dA")->M4
 #plot area sizes as color code
plot3qs(updateX(M4$A,M4$grd,M4$bas)$X,M4$grd,M4$dA)
\dontshow{\}) # examplesIf}
}
