% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{PlotA}
\alias{PlotA}
\title{PlotA}
\usage{
PlotA(M, bar = TRUE, scale_up = TRUE, ...)
}
\arguments{
\item{M}{membrane to plot M$A from}

\item{bar}{(=TRUE)if L-levels should be color-plotted}

\item{scale_up}{(=TRUE) plot A scaled-up by sqrt(bas$G.tk)}
}
\description{
plot amplitudes in matplot
}
\examples{
data("M4");SetParams(M4)
plotA(M4,ylab=expression(A[scaled]))
plot(M4,scaled_up=FALSE,ylab=expression(A[unscaled]))

}
