% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{PlotPNEM}
\alias{PlotPNEM}
\title{PlotPNEM}
\usage{
PlotPNEM(M, from = 1, to = length(M$C_PNEM))
}
\arguments{
\item{M}{membrane object to plot energies from}
}
\description{
plot recorded energies from PNEM runs
}
\examples{
data("M4")
PNEM(M4,1000)->M
PNEMVM(M,1000)->M1
PlotPNEM(M1)
}
