% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{PlotRef}
\alias{PlotRef}
\title{PlotRef}
\usage{
PlotRef(M, ARef = M$ARef)
}
\arguments{
\item{M}{membrane object to plot SEN reference shape from}

\item{ARef}{optional coefficients, eg if the objects ARef is not working}
}
\value{
-
}
\description{
plot SEN reference shape from membrane object
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
data("M4",package = "MemRBC",envir=environment())
PlotRef(M4)
\dontshow{\}) # examplesIf}
}
