% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{PlotSample}
\alias{PlotSample}
\title{PlotSample}
\usage{
PlotSample(M, last = dim(M$Sample)[1], title = "MMC sample plot", ...)
}
\arguments{
\item{M}{membrane object with $Sample keeping the MMC recorded data}

\item{last}{: integer, how many data from the tail should be plotted}

\item{title}{title of the plot, placed in separate box}

\item{...}{further plot parameters}
}
\value{
-
}
\description{
plot the curvature-energy-data sampled from MMC.
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
data("M4",package = "MemRBC",envir=environment())
MMC(M4,1000)->M4mmc
PlotSample(M4mmc)
\dontshow{\}) # examplesIf}
}
