% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PNEM.R
\name{PNEM}
\alias{PNEM}
\title{Penalized Newton Equation of Motion}
\usage{
PNEM(
  M,
  nsteps = 100,
  dt = 5e-04,
  LAfreq = 100,
  plt = TRUE,
  rho = 1,
  sd0 = 0.001,
  pltfreq = 25,
  ncores = 4,
  E_crit = 10,
  new_Av = FALSE,
  viscosity = 20,
  zero_Av = FALSE,
  mass_update_freq = 99
)
}
\arguments{
\item{M}{The input membrane with initial data and reference}

\item{nsteps}{number of time steps to be run}

\item{dt}{time step}

\item{LAfreq}{storage frequency into list of coefficients LA}

\item{plt}{(boolean) for plotting}

\item{rho}{area mass density}

\item{sd0}{standard deviation of Gaussian velocity initialization}

\item{pltfreq}{(=25) plot frequency}

\item{ncores}{(=4) for parallel threads in SCM gradient and energy}

\item{E_crit}{(=10) 1/M.Es scaled energy threshold to exit}

\item{new_Av}{set velocity according to sd0 random numbers}

\item{viscosity}{viscosity constant for damping oscillations to zero}

\item{zero_Av}{set velocity zero for pure minimization}

\item{mass_update_freq}{(=99) update frequency for mass matrix}
}
\value{
membrane object with updates from MMC with data:

LA: list of recorded coefficients A

A: last coefficients

C: last curvature

E: last total energy

Av: last coefficient velocity, good for restart

E_kin_PNEM: kinetic energy of all previous PNEMs

E_total_PNEM: total energy of all previous PNEMs

PNEMiter; total PNEM step counter, incl. previous calls

history: history of App-calls that created the result
}
\description{
run PNEM on an existing membrane object to simulate dynamics with viscosity
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
data(M4,package = "MemRBC")
SetParams(M4)
M.Rcpp=TRUE
M.Rcpp_ncores=4
M <- PNEM(M4, nsteps=10000, dt=1e-3)
plot(M)
# further data stored as attributes to coefficients A in LA
attributes(M$A)
\dontshow{\}) # examplesIf}
}
