% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PNEMVM.R
\name{PNEMVM}
\alias{PNEMVM}
\title{Penalized Newton Equation of Motion Viscosity Matrix}
\usage{
PNEMVM(
  M,
  nsteps = 100,
  dt = 5e-04,
  LAfreq = 50,
  plt = TRUE,
  filter_delta = ID,
  rho = 1,
  sd0 = 0.001,
  pltfreq = 25,
  ncores = 4,
  new_Av = FALSE,
  visc_fac = 1,
  visc_matrix = diag(sqrt(M$bas$G.tk[1]/M$bas$G.tk)),
  zero_Av = FALSE,
  mass_update_freq = 50
)
}
\arguments{
\item{M}{The input membrane with initial data and reference}

\item{nsteps}{number of time steps to be run}

\item{dt}{time step}

\item{LAfreq}{storage frequency into list of coefficients LA}

\item{plt}{(boolean) for plotting}

\item{filter_delta}{(=ID) name a function to modify for delta(A), eg damping; the filter gives back coeffs A, and it has (A,bas) as input parameters}

\item{rho}{area mass density (scalar)}

\item{sd0}{standard deviation of Gaussian velocity initialization}

\item{pltfreq}{(=25) plot frequency}

\item{ncores}{number of parallel cores in SCM energy and gradient}

\item{new_Av}{set velocity according to sd0 random numbers}

\item{visc_fac}{viscosity scaling factor for damping oscillations to zero}

\item{visc_matrix}{(=diag(sqrt(M$bas$G.tk[1]/M$bas$G.tk)) ) viscosity matrix of size $bas$Ai_max x $bas$Ai_max}

\item{zero_Av}{set velocity zero for pure minimization}

\item{mass_update_freq}{(=50) update frequency of mass matrix}
}
\value{
membrane object with updates from MMC with data:

LA: list of recorded coefficients A

A: last coefficients

C: last curvature

E: last total energy

Av: last coefficient velocity, good for restart

E_kin_PNEM: kinetic energy of all previous PNEMs

E_total_PNEM: total energy of all previous PNEMs

PNEMiter; total PNEM step counter, incl. previous calls

history: history of App-calls that created the result
}
\description{
run PNEMVM on an existing membrane object to simulate dynamics with spectral natrix viscosity
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
data(M4)
M<-M4
plot(M)
MemRBC_env$M.Rcpp=TRUE
MemRBC_env$M.Rcpp_ncores=4
M <- PNEMVM(M, nsteps=10000, dt=1e-6)
plot(M)
M
# further data stored as attributes to coefficients A in LA
attributes(last(M$LA))
\dontshow{\}) # examplesIf}
}
