% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSD.R
\name{PSD}
\alias{PSD}
\title{Penalized Steepest Descend}
\usage{
PSD(
  M,
  nsteps = 100,
  del = 1e-06,
  plt = FALSE,
  pltfreq = 10,
  LAfreq = 100,
  ncores = 4,
  filter_strength = 0
)
}
\arguments{
\item{M}{The input membrane with initial data and reference}

\item{nsteps}{(=100) number of PSDC steps to be performed}

\item{del}{(=1e-6) step-size}

\item{plt}{(=FALSE) for plotting}

\item{pltfreq}{(=10) for plotting every pltfreq step}

\item{LAfreq}{(=100) storage frequency into list of coefficients $LA}

\item{ncores}{(=4) parallel threads in SCM energy and gradient}

\item{filter_strength}{(=0) to scale down higher-l gradient components by `Filter = (1 + filter_strength * 1/sqrt(M$bas$G.tk))`}
}
\value{
membrane object with updates from MMC with data:

LA: list of recorded coefficients A

A: last coefficients

C: last curvature

E: last total energy

E_PSD: all recorded total energies

PSDiter; total PSD steps, incl. previous calls

history: history of App-calls that created the result
}
\description{
run PSD on an existing membrane object to reduce energy by gradient of energy.
A more efficient minimizer may be the PNEM.
}
\examples{
\dontshow{if (exists("L_Ylm")) withAutoprint(\{ # examplesIf}
data("M4",package = "MemRBC")
M<-M4
plot(M,alpha=0.65)
M <- PSD(M,nsteps=100,del=1e-6,LAfreq=10)
plot(M,add=TRUE,col="red")
M
attributes(last(M$LA))
\dontshow{\}) # examplesIf}
}
