% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{Quantities}
\alias{Quantities}
\title{Quantities}
\usage{
Quantities(M)
}
\arguments{
\item{M}{The input membrane with initial data and reference}
}
\value{
quantities Area, Volume, curvature Curv and bending energy Wb
}
\description{
report some quantities from a membrane object
}
\examples{
data("M1")
Quantities(M1)
}
