% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{ReduceM1}
\alias{ReduceM1}
\title{ReduceM1}
\usage{
ReduceM1(M)
}
\arguments{
\item{M}{membrane object to change}
}
\value{
modified M with reduced basis, but $Ref unchanged.
}
\description{
Reduce basis in M to leading cos v, sin v terms, i.e. |m|<2 in Y_lm.
Requires update(M,"Ref"), if SEN is used (i.e. M.Ka and/or M.mu not zero.
It is recommended to save a reduced model with save_MemRBC(...,reduce_basis=FALSE),
to keep the basis in reduced form in the file.
For loading, you must then use load_MemRBC(..., unreduce_basis=FALSE) to not overwrite the reduced basis.
}
\examples{
data("US_L10_de_novo"); SetParams(US_L10_de_novo)
ReduceM1(US_L10_de_novo) -> L10fast
# should reduce from 120 to 30 rows in bas$A
#
# not needed, since paraneters switch off SEN::
update(L10,what="Ref")->L10fast
plot(L10fast)

MMC(L10fast, 20000)->M10fast_mmc
PlotSample(L10fast_mmc,last=20000)

}
