% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{Rewind}
\alias{Rewind}
\title{Rewind to set coefficient M$A by n records back from end of list of recorded coeffs M$LA
- practically you do the same with
M$A=M$LA[[length(M$LA)-n]]
ATTENTION: while M$A usually holds the lasrt coefficients from an App
 last entries in LA hold older coeffs, in particular if LAfreq>1 in the last App call}
\usage{
Rewind(M, n = 1)
}
\arguments{
\item{M}{membrane object with list $LA of recorded coefficients}

\item{n}{(=1) how many records to rewind?}
}
\value{
membrane object with M$A set to `LA[[lengh(LA)-n]]`
}
\description{
Rewind to set coefficient M$A by n records back from end of list of recorded coeffs M$LA
- practically you do the same with
M$A=M$LA[[length(M$LA)-n]]
ATTENTION: while M$A usually holds the lasrt coefficients from an App
 last entries in LA hold older coeffs, in particular if LAfreq>1 in the last App call
}
