% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{Rewind}
\alias{Rewind}
\title{Rewind}
\usage{
Rewind(M, last = 2)
}
\arguments{
\item{M}{membrane object with list $LA of recorded coefficients}

\item{n}{(=1) how many records to rewind?}
}
\value{
membrane object with M$A set to LA[[lengh(LA)-n]]
}
\description{
rewind coefficients in M$A by n records from the list of recorded coeffs M$LA
 - practically you do the same with
 M$A=M$LA[[length(M$LA)-n]]
}
