% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{RosenProjection}
\alias{RosenProjection}
\title{RosenProjection}
\usage{
RosenProjection(G, g2, bas)
}
\arguments{
\item{G}{input gradient}

\item{g2}{current energy}

\item{bas}{basis}
}
\description{
Rosen (1961) projection of gradient
}
