% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotUV.R
\name{rotUV}
\alias{rotUV}
\title{rotUV}
\usage{
rotUV(Min, du, dv, plt = FALSE, transpose = FALSE)
}
\arguments{
\item{du}{angles of rotation relative to Z-axis}

\item{dv}{angle of rotation around Z-axis}

\item{transpose}{(=FALSE) for backward rotation, not verified}

\item{M}{membrane object to rotate}
}
\value{
membrane object with coefficients after rotation, but original SEN reference
}
\description{
rotate membrane coordinates and re-compute coefficients
without modifying the SEN reference.
rotUV may help to move poles to Z-axis or for more homogeneous distribution of spatial points.
}
\examples{
data_MemRBC("M_stomatocyte_L12")
plot(M_stomatocyte_L12)
rotUV(M_stomatocyte_L12,pi/2,3*pi/4)->M
plot(M)
}
