% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{SE_2_OBJ}
\alias{SE_2_OBJ}
\title{SE_2_OBJ}
\usage{
SE_2_OBJ(f.in, f.out, comment = "o")
}
\arguments{
\item{f.in, f.out}{: names of input SE dump /output .obj file}

\item{comment}{(="o") how Evolver input file encodes comments}
}
\value{
3d mesh object
}
\description{
Read data from surface evolver dump file.
The vertices are renumbered here, taking
more computing time.
}
