% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{SErenumbered_2_OBJ}
\alias{SErenumbered_2_OBJ}
\title{SErenumbered_2_OBJ}
\usage{
SErenumbered_2_OBJ(f.in, f.out, comment = "o")
}
\description{
Read data from surface evolver dump file
 special case: faster if SE renumbered all before the dump
}
