% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{set_A_to_Ref}
\alias{set_A_to_Ref}
\title{set_A_to_Ref}
\usage{
set_A_to_Ref(M)
}
\arguments{
\item{M}{membrane object to copy coefficients into}
}
\value{
modified M2 with new coefficients from M$ARef; spectral order remains unchanged
}
\description{
copy reference shape coefficients $ARef to current coeffs $A.
}
