% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{SetParams}
\alias{SetParams}
\title{SetParams}
\usage{
SetParams(M)
}
\arguments{
\item{M}{Object to take parameters from (stored in M$Params)}
}
\value{
-
}
\description{
Set parameters from membrane object to MemRBC_env variables
 since data(...) cannot load multiple global data, after data() you should call SetParams on the loaded object.
}
