% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{StoreParams}
\alias{StoreParams}
\title{StoreParams}
\usage{
StoreParams(M)
}
\arguments{
\item{M}{Membrane to store actual parameters like MemRBC_env$M.0, M.K_ADE etc. into}
}
\value{
MemRBC object, with Params from `MemRBC_env`
}
\description{
Store current MemRB_env membrane parameters in $Param of the object; useful for load and save
}
