% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{transplant}
\alias{transplant}
\title{transplant}
\usage{
transplant(M1, M2, plt = FALSE)
}
\arguments{
\item{M1}{membrane object to copy coefficients from}

\item{M2}{membrane object to copy coefficients into}
}
\value{
modified M2 with new coefficients
}
\description{
copy coefficients from one membrane to another
}
\examples{
data("M4")
data("S5")
transplant(S5,M4),S4mod
S4mod
}
