% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{TriMesh_Unduloid}
\alias{TriMesh_Unduloid}
\title{TriMesh_Unduloid}
\usage{
TriMesh_Unduloid(
  a = 1,
  c = 2,
  periods = 1,
  nx = 40,
  ny = 40,
  shade = TRUE,
  wire = FALSE,
  clean = TRUE
)
}
\description{
create an Unduloid 3d object for a fraction or  multiple periods.
Unduloids may be interesting shapes to fit, see example.
}
\examples{
# a special grid is made with hole at north and south pole,
# ie, u starts not at zero
open=0.4 # also higher work, but objects boundary is never fitted
g<-MakeGrid_GaussLegendreSimpson(180,ua=open,ub=pi-open)
g$ndof
range(g$U)
(U<-TriMesh_Unduloid(periods=4,nx=g$nu,ny=g$nv-1, a=1,c=0.1,clean=FALSE))
attr(U,"H_theor") # theoretical mean curvature
attr(U,"H_vcg_6") # mean from vertices with 6 neighbors
#  mean curvature for other vertices is problematic in vcg
b<-MakeBasis_UV(23,g$U,g$V)
# one should exclude double coordinates for
# the fit, so mask is needed
b$mask<-double_uv_ind(b$uv[,1],b$uv[,2])
CenterX(Obj2X(U)) -> X
cat(dim(X)[1],"?=", g$ndof,"\n")
if (dim(X)[1] == g$ndof)
{ # if not matching repeat with alternative n in grid
rgl::plot3d(X,col=2,aspect=FALSE)
X2Obj(U,X) -> U1
rgl::contourLines3d(U1,b$uv[,1],levels=(0:100)*pi/100)
rgl::contourLines3d(U1,b$uv[,2],40)
rgl::shade3d(U1,col="grey",alpha=0.5)
A<-FitAlm_Tikhonov(X,b,lambda=0) # , WX=sin(g$U))
A[,3]<--A[,3] # wrong orientation correction
# you may try the fit with weights, WX=sin(g$U)
MakeMemRBC(A,g,b)->M
rgl::open3d()
rgl::plot3d(X,aspect=FALSE,alpha=0.45)
plot(M,alpha=0.45,col="cyan",wire=FALSE)
E=E_SCM(M$A,M$grd,M$bas,updateX(M$A,M$grd,M$bas))
-E$Curv/E$Area/2 # mean curvature from integral over area
attr(U,"H_vcg_6") # comparison with original
imag.obj.colorbar(U1,E$curv)
rgl::title3d("curvature density")
X1<-updateX(M$A,M$grd,M$bas)$X
M$grd$Obj<-X2Obj(M$grd$Obj,X1)
rgl::shade3d(M$grd$Obj,alpha=0.2)
mean(E$curv/E$dA/2)
M.C0=0;M.mu=0;M.Ka=0
M$bas$Target[1:2]=c(E$Area,E$Volume)
save_MemRBC(M,"Unduloid.rdat")
MMC(M,1000,plt=TRUE,pltfreq=2,C0=0)
}
}
