% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{two_draw3d}
\alias{two_draw3d}
\title{two_draw3d}
\usage{
two_draw3d(A, M, cont = FALSE, title = "", x = 400, y = 400)
}
\arguments{
\item{A}{coefficients to use}

\item{M}{the input membrane to be plotted}

\item{cont}{(=FALSE) draw gridlines u=const, v=const}
}
\value{
-
}
\description{
plot membrane object with area stress alpha and shear beta as color
}
\examples{
data("M_stomatocyte_L12"); M<-M_stomatocyte_L12
 perturb original coefficients
A <- pertA_Gauss(M$A, M$bas, sd=0.2)
two_screens(); two_draw3d(A, M)
}
