% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{two_screens3d}
\alias{two_screens3d}
\title{two_screens3d}
\usage{
two_screens3d(x = 400, y = 400)
}
\arguments{
\item{x, y}{dimensions in pixels}
}
\value{
variables MemRBC_env$M.scr1 and MemRBC_env$M.scr2 are set to device ids
}
\description{
open two screens for 3D-plots by two_draw3d()
}
\examples{
two_screens3d(x=650,y=300)
}
