% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC-Apps.R
\name{update.MemRBC}
\alias{update.MemRBC}
\title{update}
\usage{
\method{update}{MemRBC}(M, what = c("dA", "Quantities", "X"), n = 0, L = 5)
}
\arguments{
\item{M}{the input membrane to be updated}

\item{what}{vector of character from "dA", "Quantities", "Basis", "Grid", "Ref", "curv", "SCM", "X", "SEN"}

\item{n}{for what="Grid": new number of grid points}

\item{L}{for what="Basis": spectral order L}
}
\value{
updated membrane object
}
\description{
update data in membrane object
}
\examples{
data("M_stomatocyte_L12")
update(M_stomatocyte_L12,"X")->M
plot3d(M$X, aspect=FALSE)
# make lower spectral order membrane from M
update(M, what=c("Grid","Basis","Ref"), n=30, L=8)->L8
two_screens3d(); two_draw3d(L8$A, L8)
rgl::open3d(); plot(L8)
Quantities(L8)
Energy(L8)
}
