% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembraneRBC.R
\name{updateX_only}
\alias{updateX_only}
\title{updateX_only}
\usage{
updateX_only(A, grd, bas)
}
\arguments{
\item{A}{coefficients of shape}

\item{grd}{grid from on which the basis is computed}

\item{bas}{basis function values $Ylm}
}
\value{
basis object, type is list.
}
\description{
updates coordinates C$X, NOT the derivatives.
}
